% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_dags.R
\name{add_edges}
\alias{add_edges}
\title{Add edges to a DAG}
\usage{
add_edges(parent, children)
}
\arguments{
\item{parent}{Character. The name of the parent variable.}

\item{chilren}{Character scalar or vector. The name(s) of the child(ren) variable(s).}
}
\value{
Edge(s) in a DAG
}
\description{
Add edges to a DAG
}
\examples{

dag1 <- make_dag(
add_edges(parent = "X",children = c("K","Y")),
add_edges(parent = "K",children = "Y")
)

}
