% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{add_prior_distribution}
\alias{add_prior_distribution}
\title{Add prior distribution draws}
\usage{
add_prior_distribution(model, n_draws = 4000)
}
\arguments{
\item{model}{A model generated by \code{make_model}}

\item{n_draws}{Number of draws}
}
\description{
Add n_param x n_draws database of possible lambda draws to the model
}
