% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type_helpers.R
\name{add_wildcard}
\alias{add_wildcard}
\title{#' Adds a wildcard for every missing parent}
\usage{
add_wildcard(node, statement, parents, missing_parents)
}
\arguments{
\item{node}{A quoted name of a node}

\item{statement}{A quoted causal statement.}
}
\description{
#' Adds a wildcard for every missing parent
}
\keyword{internal}
