% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{adjust_edge}
\alias{adjust_edge}
\title{internal helper
adjust_edge moves the base and tip of edge arrows so as to avoid overlap
with nodes}
\usage{
adjust_edge(dag, nodesize, extent)
}
\description{
internal helper
adjust_edge moves the base and tip of edge arrows so as to avoid overlap
with nodes
}
\keyword{internal}
