% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_strategies.R
\name{all_possible}
\alias{all_possible}
\title{helper for getting all data on variables with N observed}
\usage{
all_possible(model, N, vars = NULL, condition = TRUE)
}
\description{
helper for getting all data on variables with N observed
}
\examples{
model <- make_model("X->M->Y")
gbiqq:::all_possible(model, N=2, vars = c("X", "M"))
gbiqq:::all_possible(model, N=2, vars = c("X", "Y"), condition = "Y==0")
}
