% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_strategies.R
\name{allocations}
\alias{allocations}
\title{helper for ways to allocate N units into n data types}
\usage{
allocations(N, n)
}
\arguments{
\item{N}{Number of observations to be distributed}

\item{n}{Number of possible values observations could take}
}
\description{
helper for ways to allocate N units into n data types
}
\examples{
allocations(4,2)
}
