% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{any_discrepancies}
\alias{any_discrepancies}
\title{Check for discrepancies
Used in make_alphas}
\usage{
any_discrepancies(alphas, translated_alphas,
  query_alpha = attr(translated_alphas, "query"))
}
\description{
Check for discrepancies
Used in make_alphas
}
