% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{attractive_force}
\alias{attractive_force}
\title{internal helper
calculates the attractive force component used in generating
force-directed graph layouts}
\usage{
attractive_force(distance, desired_distance, k)
}
\description{
internal helper
calculates the attractive force component used in generating
force-directed graph layouts
}
\keyword{internal}
