% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_and_estimands.R
\name{calculate_multiple_estimands}
\alias{calculate_multiple_estimands}
\title{Calculate multipl estimands}
\usage{
calculate_multiple_estimands(model = make_model("X" \%->\% "Y"),
  posteriors = list(FALSE), dos = list(op_1 = list(do_1 = list(X = 1),
  do_2 = list(X = 0))), qs = list(q1 = function(Q1, Q2) Q1$Y == 1 & Q2$Y
  == 0), subsets = list(TRUE), estimand_labels = NULL, stats = c(mean
  = mean, sd = sd))
}
\arguments{
\item{model}{A  model}

\item{posterior}{if true use a posterior distribution, otherwise use the prior}

\item{subset}{quoted expression evaluates to logical statement. subset allows estimand to be conditioned on *observational* distribution.}
}
\description{
Calculated from a prior or posterior distribution
}
\examples{
model <- make_model("X" \%->\% "Y")
model <- add_prior_distribution(model)
calculate_multiple_estimands(model,
           posteriors = FALSE,
           dos = list(op_1 =  list(do_1 = list(X = 1), do_2 = list(X = 0)),
										 op_2 =  list(do_1 = list(X = 1), do_2 = list(X = 0))),
					qs =  list(q1   =  function(Q1, Q2) Q1$Y -  Q2$Y,
										 q2   =  function(Q1, Q2) Q1$Y == 1 & Q2$Y == 0),
					subsets = TRUE,
					estimand_labels = c("ATE", "Share_positive"))

}
