% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalQueries-package.R
\docType{package}
\name{CausalQueries-package}
\alias{CausalQueries}
\alias{CausalQueries-package}
\title{CausalQueries: Make, Update, and Query Binary Causal Models}
\description{
'CausalQueries' is a package that lets users generate binary causal models,
update over models given data, and calculate arbitrary causal queries.
Model definition uses arrow syntax for causal statements
(e.g. \code{"X -> Y"}, \code{"X <-> Y"}).
Updating is implemented in 'stan'.
}
\references{
Tietz T, Medina L, Syunyaev G, Humphreys M (2026).
"Making, Updating, and Querying Causal Models with CausalQueries."
\emph{Journal of Statistical Software}, \bold{117}(1), 1--40.
\doi{10.18637/jss.v117.i01}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://integrated-inferences.github.io/CausalQueries/}
  \item Report bugs at \url{https://github.com/integrated-inferences/CausalQueries/issues}
}

}
\author{
\strong{Maintainer}: Till Tietz \email{ttietz2014@gmail.com} (\href{https://orcid.org/0000-0002-2916-9059}{ORCID})

Authors:
\itemize{
  \item Till Tietz \email{ttietz2014@gmail.com} (\href{https://orcid.org/0000-0002-2916-9059}{ORCID})
  \item Macartan Humphreys \email{macartan@gmail.com} (\href{https://orcid.org/0000-0001-7029-2326}{ORCID})
  \item Alan Jacobs \email{alan.jacobs@ubc.ca}
  \item Lily Medina \email{lilymiru@gmail.com} (\href{https://orcid.org/0009-0004-2423-524X}{ORCID})
  \item Georgiy Syunyaev \email{georgiy.syunyaev@vanderbilt.edu} (\href{https://orcid.org/0000-0002-4391-6313}{ORCID})
}

Other contributors:
\itemize{
  \item Clara Bicalho \email{clarabmcorreia@gmail.com} [contributor]
  \item Jasper Cooper \email{jjc2247@columbia.edu} [contributor]
  \item Merlin Heidemanns \email{mnh2123@columbia.edu} [contributor]
  \item Julio Solis \email{juliosolisar@gmail.com} [contributor]
  \item Moritz Marbach \email{m.marbach@ucl.ac.uk} (\href{https://orcid.org/0000-0002-7101-2821}{ORCID}) [contributor]
}

}
