% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_model.R
\name{CausalQueries}
\alias{CausalQueries}
\title{CausalQueries}
\usage{
CausalQueries(...)
}
\arguments{
\item{...}{arguments passed to  \code{\link{update_model}}}
}
\description{
An alias for \code{\link{update_model}}
}
