% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_model.R
\name{check_args}
\alias{check_args}
\title{helper to check arguments}
\usage{
check_args(model, using, given, queries, case_level, fun)
}
\arguments{
\item{model}{passed from parent function}

\item{using}{passed from parent function}

\item{given}{passed from parent function}

\item{fun}{string specifying the name of the parent function}

\item{query}{passed from parent function}
}
\value{
list of altered arguments
}
\description{
helper to check arguments
}
\keyword{internal}
