% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_query}
\alias{check_query}
\title{Warn about improper query specification and apply fixes}
\usage{
check_query(query)
}
\arguments{
\item{query}{a string specifying a query}
}
\value{
fixed query as string
}
\description{
Warn about improper query specification and apply fixes
}
\keyword{internal}
