% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_helpers.R
\name{check_string_input}
\alias{check_string_input}
\title{Check string_input}
\usage{
check_string_input(param_list = list(), call_name = NULL)
}
\arguments{
\item{param_list}{List of parameters}

\item{call_name}{Name of the call.}
}
\description{
Check string_input
}
