% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chickering_pearl}
\alias{chickering_pearl}
\title{Chickering Pearl Data}
\format{
A data frame with 8 rows and 3 columns:
\describe{
  \item{event}{The data type}
  \item{strategy}{For which nodes is data available}
  \item{count}{Number of units with this data type}
}
}
\source{
\url{https://cdn.aaai.org/AAAI/1996/AAAI96-188.pdf}
}
\usage{
chickering_pearl
}
\description{
A compact dataset containing information on an encouragement, (Z, cholestyramine prescription), a treatment (X, usage), and an outcome (Y, cholesterol).
From David Maxwell Chickering and Judea Pearl: "A Clinician’s Tool for Analyzing Non-compliance", AAAI-96 Proceedings.
}
\keyword{datasets}
