% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_condition}
\alias{clean_condition}
\title{Clean condition}
\usage{
clean_condition(condition)
}
\arguments{
\item{condition}{A character string. Condition that refers to a unique position (possible outcome) in a nodal type.}
}
\description{
Takes a string specifying condition and returns properly spaced string.
}
\keyword{internal}
