% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_params.R
\name{clean_param_vector}
\alias{clean_param_vector}
\title{Clean parameter vector}
\usage{
clean_param_vector(model, parameters)
}
\arguments{
\item{model}{A model made by make_model}

\item{parameters}{A parameter vector (real numbers in [0,1])}
}
\description{
Clean parameter vector
}
\examples{
model <- make_model("X->Y")
gbiqq:::clean_param_vector(model, 1:6)
}
