% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{clean_statement}
\alias{clean_statement}
\title{Helper to clean and check the validity of causal statements specifying a DAG.
This function isolates nodes and edges specified in a causal statements and
makes them processable by \code{make_dag}}
\usage{
clean_statement(statement)
}
\arguments{
\item{statement}{character string. Statement describing causal
relations between nodes.}
}
\value{
a list of nodes and edges specified in the input statement
}
\description{
Helper to clean and check the validity of causal statements specifying a DAG.
This function isolates nodes and edges specified in a causal statements and
makes them processable by \code{make_dag}
}
\keyword{internal}
