% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_ambiguity_matrices.R
\name{collapse_ambiguity_matrices}
\alias{collapse_ambiguity_matrices}
\title{Create matrix mapping endogenous type combinations to data}
\usage{
collapse_ambiguity_matrices(dag)
}
\arguments{
\item{dag}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
test_dag <-
     gbiqq::make_dag(add_edges(parent = "X",children = c("K","Y")),
     								add_edges(parent = c("K"),children = "Y"))
gbiqq::collapse_ambiguity_matrices(test_dag)

}
