% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodal_types.R
\name{collapse_nodal_types}
\alias{collapse_nodal_types}
\title{collapse nodal types}
\usage{
collapse_nodal_types(nodal_types)
}
\arguments{
\item{nodal_types}{A list of nodal types.}
}
\description{
collapse nodal types
}
