% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{combine_lists}
\alias{combine_lists}
\title{combine two lists by names}
\usage{
combine_lists(list1, list2)
}
\arguments{
\item{list1}{a list}

\item{list2}{a list typically different than list1}
}
\description{
combine two lists by names
}
