% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{compact_to_full}
\alias{compact_to_full}
\title{Generate full dataset from compact data}
\usage{
compact_to_full(model, compact_df)
}
\arguments{
\item{model}{A model created by make_model()}

\item{compact_df}{A compact dataframe compatible with model}

\item{n}{Number of observations}
}
\description{
Generate full dataset from compact data
}
\examples{
model <- make_model(add_edges(parent = "X", children = c("Y")))
compact_df <- draw_data(model = model, n = 4)
compact_to_full(model, compact_df)
}
