% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_helpers.R
\name{complements}
\alias{complements}
\title{Make statement for complements}
\usage{
complements(X1, X2, Y)
}
\arguments{
\item{X1}{A character. The quoted name of the input node 1.}

\item{X2}{A character. The quoted name of the input node 2.}

\item{Y}{A character. The quoted name of the outcome node.}
}
\value{
A character statement of class \code{statement}
}
\description{
Generate a statement for X1, X1 complement each other in the production of Y
}
\examples{
\donttest{
complements('A', 'B', 'W')
}
}
\seealso{
Other statements: 
\code{\link{decreasing}()},
\code{\link{increasing}()},
\code{\link{interacts}()},
\code{\link{non_decreasing}()},
\code{\link{non_increasing}()},
\code{\link{substitutes}()},
\code{\link{te}()}
}
\concept{statements}
