% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{confound_curve_curvatures}
\alias{confound_curve_curvatures}
\title{Signed geom_curve curvatures for confound edges.}
\usage{
confound_curve_curvatures(dag, pos, strength, confound_bulge)
}
\description{
Signed geom_curve curvatures for confound edges.
}
\keyword{internal}
