% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_par_values.R
\name{construct_commands_alter_at}
\alias{construct_commands_alter_at}
\title{make_par_values}
\usage{
construct_commands_alter_at(alter_at)
}
\arguments{
\item{alter_at}{string specifying filtering operations to be applied to
parameters_df, yielding a logical vector indicating parameters for which
values should be altered.}
}
\value{
string specifying a filter command
}
\description{
helper to generate filter commands specifying rows of parameters_df that
should be altered given an alter_at statement
}
\keyword{internal}
