% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_par_values.R
\name{construct_commands_param_names}
\alias{construct_commands_param_names}
\title{make_par_values}
\usage{
construct_commands_param_names(param_names, model_param_names)
}
\arguments{
\item{param_names}{vector of strings. The name of specific parameter in
the form of, for example, 'X.1', 'Y.01'}

\item{model_param_names}{vector of strings. Parameter names found
in the model.}
}
\value{
string specifying a filter command
}
\description{
helper to generate filter commands specifying rows of parameters_df that
should be altered given an a vector of parameter names
}
\keyword{internal}
