% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_confounds.R
\name{continue_names}
\alias{continue_names}
\title{Continue names}
\usage{
continue_names(x)
}
\arguments{
\item{x}{A vector with strings of the form c("x-1", "x-2")}
}
\description{
Slightly hacky function to continue param_set names in a sequence
}
\examples{
x <- c("S-3", "S-3", "S-5")
gbiqq:::continue_names(x)
}
