% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{curve_segment_df}
\alias{curve_segment_df}
\title{Segment endpoints for ggplot2::geom_curve layers.}
\usage{
curve_segment_df(dag, pos, kind)
}
\description{
Segment endpoints for ggplot2::geom_curve layers.
}
\keyword{internal}
