% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{data_strategy}
\alias{data_strategy}
\title{Data Strategy}
\usage{
data_strategy(model, parameters = NULL, n_obs = NULL, n = NULL,
  vars = list(NULL), probs = list(NULL), subsets = list(NULL),
  using = "priors")
}
\arguments{
\item{using}{String indicating whether to use `priors`, `posteriors` or `parameters`}
}
\description{
Data Strategy
}
\examples{
# A strategy in which X, Y are observed for sure and M is observed with 50\% probability for X=1, Y=0 cases
model <- make_model("X -> M -> Y")
model <- set_parameters(model, type = "flat")
data_strategy(
  model,
  n_obs = 8,
  n = NULL,
  vars = list(c("X", "Y"), "M"),
  probs = list(1, .5),
  subsets = list(NULL, "X==1 & Y==0"))
}
