% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmap.R
\name{data_to_data}
\alias{data_to_data}
\title{helper to generate a matrix mapping from names of M to names of A}
\usage{
data_to_data(M, A)
}
\description{
helper to generate a matrix mapping from names of M to names of A
}
