% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_helpers.R
\name{decreasing}
\alias{decreasing}
\title{Make monotonicity statement (negative)}
\usage{
decreasing(X, Y)
}
\arguments{
\item{X}{input node}

\item{Y}{outcome node}
}
\value{
A character statement of class statement
}
\description{
Generate a statement for Y monotonic (decreasing) in X
}
\examples{
decreasing('A', 'B')

}
\seealso{
Other statements: 
\code{\link{complements}()},
\code{\link{increasing}()},
\code{\link{interacts}()},
\code{\link{non_decreasing}()},
\code{\link{non_increasing}()},
\code{\link{substitutes}()},
\code{\link{te}()}
}
\concept{statements}
