% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{default_stan_control}
\alias{default_stan_control}
\title{default_stan_control}
\usage{
default_stan_control(adapt_delta = NULL, max_treedepth = 15L)
}
\description{
default_stan_control
}
\details{
Sets controls to default unless otherwise specified.
}
