% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan_using_dags.R
\name{define_priors}
\alias{define_priors}
\title{Define priors}
\usage{
define_priors(dag = NULL, nodal_types = NULL,
  prior_distribution = NULL)
}
\arguments{
\item{dag}{A dag created by make_dag()}

\item{nodal_types}{}
}
\description{
Define priors
}
