% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{democracy_data}
\alias{democracy_data}
\title{Development and Democratization: Data for replication of analysis in *Integrated Inferences*}
\format{
A data frame with 84 rows and 5 nodes:
\describe{
  \item{Case}{Case}
  \item{D}{Democracy}
  \item{I}{Inequality}
  \item{P}{International Pressure}
  \item{M}{Mobilization}
}
}
\source{
\url{https://www.cambridge.org/core/journals/american-political-science-review/article/inequality-and-regime-change-democratic-transitions-and-the-stability-of-democratic-rule/C39AAF4CF274445555FF41F7CC896AE3#fndtn-supplementary-materials/}
}
\usage{
democracy_data
}
\description{
A dataset containing information on inequality, democracy, mobilization, and international pressure.
Made by \code{devtools::use_data(democracy_data, CausalQueries)}
}
\keyword{datasets}
