% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_model.R
\name{deparse_given}
\alias{deparse_given}
\title{helper to separate query and givens in query statement}
\usage{
deparse_given(query)
}
\description{
helper to separate query and givens in query statement
}
