% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{get_ambiguities}
\alias{make_ambiguity_matrix_internal}
\alias{make_ambiguity_matrices}
\alias{map_types_to_data}
\alias{get_pi_expanders}
\alias{get_chains}
\alias{get_n_endogenous_types}
\alias{get_children_of_exogenous}
\alias{get_terminal_vars}
\alias{get_ancestors}
\alias{get_exogenous_vars}
\alias{get_endogenous_vars}
\alias{get_variables}
\title{Higher level function that returns all of the type ambiguities for all possible patterns of evidence}
\usage{
get_ambiguities(model)

make_ambiguity_matrix_internal(ambiguities)

make_ambiguity_matrices(model)

map_types_to_data(model)

get_pi_expanders(pi, model)

get_chains(model)

get_n_endogenous_types(model)

get_children_of_exogenous(model)

get_terminal_vars(model)

get_ancestors(model)

get_exogenous_vars(model)

get_endogenous_vars(model)

get_variables(model)
}
\arguments{
\item{model}{A dag created by make_dag()}

\item{ambiguities}{Ambiguities matrix produced by \code{get_ambiguities()}}

\item{dag}{A dag created by make_dag()}

\item{model}{A dag created by make_dag()}

\item{model}{A dag created by make_dag()}

\item{model}{DESCRIPTION.}

\item{A}{probabilistic causal model created by make_model()}

\item{A}{probabilistic causal model created by make_model()}

\item{dag}{DESCRIPTION.}

\item{A}{probabilistic causal model created by make_model()}

\item{A}{probabilistic causal model created by make_model()}

\item{A}{probabilistic causal model created by make_model()}

\item{A}{probabilistic causal model created by make_model()}
}
\value{
A list of mappings between types

A list of mappings between types

A list of mappings between types and data

A list of pi helpers

RETURN_DESCRIPTION

A list of the numbers of endogenous types

A list of the children of exogenous variables

RETURN_DESCRIPTION

A list of ancestors in a DAG

A vector of exogenous variables

A vector of endogenous variables

A vector of variable names
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
test_dag <-
make_model(add_edges(parent = "X",children = c("K", "Y1")),
								add_edges(parent = "K", children = c("Y1","Y2")),
								add_edges(parent = "Z", children = c("K")))

get_chains(test_dag)
}

\dontrun{
dag <- make_dag(
 add_edges(parent = "X",children = c("K","Y")),
 add_edges(parent = "K",children = "Y")
)

get_ancestors(dag)
}
}
