% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_strategies.R
\name{diagnose_strategies_single}
\alias{diagnose_strategies_single}
\title{Diagnosis a single data strategy}
\usage{
diagnose_strategies_single(reference_model, analysis_model, given, queries,
  estimates_database = NULL, possible_data = NULL,
  possible_data_args = list(N = list(1), within = FALSE, condition =
  list(TRUE), vars = list(NULL)), sims = 1000)
}
\description{
Diagnosis a single data strategy
}
