% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_event_prob.R
\name{draw_event_prob}
\alias{draw_event_prob}
\title{Draw event probabilities}
\usage{
draw_event_prob(model, P = NULL, A = NULL, parameters = NULL,
  type_prob = NULL, using = NULL)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{P}{A matrix. Parameter matrix. Not required but may be provided to avoid repeated computation for simulations.}

\item{A}{A matrix. Ambiguity matrix. Not required but may be provided to avoid repeated computation for simulations.}

\item{parameters}{A numeric vector. Values of parameters may be specified. By default, parameters is drawn from priors.}

\item{type_prob}{A numeric vector. Type probabilities (usually not required).}

\item{using}{String indicating whether to use "priors", "posteriors" or "parameters".}
}
\description{
Draw event probabilities
}
\examples{
model <- make_model("X -> Y")
draw_event_prob(model = model)
draw_event_prob(model = model, parameters = rep(1, 6))
draw_event_prob(model = model, parameters = 1:6)
draw_event_prob(model = model, using = "priors")
draw_event_prob(model = model, using = "parameters")
}
