% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_type_prob}
\alias{draw_type_prob}
\title{Draw type probabilities}
\usage{
draw_type_prob(model, P = NULL, parameters = NULL, using = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{P}{Parameter matrix. Not required but can avoid repeated computation for simulations}

\item{parameters}{A specific parameter vector may be provided, otherwise parameters are drawn using `using` and `draw_parameters`}

\item{using}{String indicating whether to use `priors`, `posteriors` or `parameters`}
}
\description{
Draw type probabilities
}
\examples{
model <- make_model("X -> Y")
draw_type_prob(model = model)
}
