% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_type_prob_multiple}
\alias{draw_type_prob_multiple}
\title{Draw matrix of type probabilities, before or after estimation}
\usage{
draw_type_prob_multiple(model, posterior = FALSE, n_draws = 4000)
}
\arguments{
\item{model}{A model, normally containing a prior or a posterior distribution}

\item{posterior}{if true use a posterior distribution, otherwise use the prior}

\item{n_draws}{If no prior distribution provided, generate prior distribution with n_draws draws}
}
\description{
Draw matrix of type probabilities, before or after estimation
}
