% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{drop_empty_families}
\alias{drop_empty_families}
\title{Drop empty families}
\usage{
drop_empty_families(data_events)
}
\arguments{
\item{data_events}{A dataframe whose columns are \code{event}, \code{strategy} and \code{count}}
}
\value{
Returns data events with strategies (excluding  strategy families that contain no observed data)
}
\description{
Drop empty families
}
\examples{

data_events <- data.frame(event = c("X0Y0", "Y0"), strategy = c("XY", "Y"), count = 1:0)
gbiqq:::drop_empty_families(data_events)

}
