% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{edges_intersect}
\alias{edges_intersect}
\title{internal helper
checks whether two edges intersect given the edge start and endpoints
this is used to apply penalties for edge crossing in the force directed
graph layout}
\usage{
edges_intersect(p1, p2, p3, p4)
}
\description{
internal helper
checks whether two edges intersect given the edge start and endpoints
this is used to apply penalties for edge crossing in the force directed
graph layout
}
\keyword{internal}
