% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{encode_data}
\alias{encode_data}
\title{Encode data}
\usage{
encode_data(model, data)
}
\arguments{
\item{model}{A  model}

\item{data}{Data in long format}
}
\description{
Takes data in long format, including NA values or blanks and returns vector with each row encoded as a data type.
}
\examples{
model <- make_model("X -> Y")
data <- simulate_data(model, n = 4)
data[1,1] <- ""
data[3,2] <- NA

encode_data(model, data)
}
