% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{exclude_node_from_causal_type}
\alias{exclude_node_from_causal_type}
\title{Exclude node from causal type restriction
Returns updated vector of rows in causal type matrix to be excluded, taking into account nodal types affected by restriction.}
\usage{
exclude_node_from_causal_type(var_name, causal_types, in_restriction)
}
\arguments{
\item{var_name}{A string. Variable name in a model.}

\item{causal_types}{A data.frame of causal types of a model.}

\item{in_restriction}{A logical vector. Rows of \code{causal_type} in restriction query.}
}
\value{
A logical vector
}
\description{
Exclude node from causal type restriction
Returns updated vector of rows in causal type matrix to be excluded, taking into account nodal types affected by restriction.
}
