% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguities.R
\name{expand_ambiguity_matrices_internal}
\alias{expand_ambiguity_matrices_internal}
\title{FUNCTION_TITLE}
\usage{
expand_ambiguity_matrices_internal(model)
}
\arguments{
\item{model}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
