% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_type.R
\name{expand_nodal_expression}
\alias{expand_nodal_expression}
\title{Helper to expand nodal expression}
\usage{
expand_nodal_expression(model, query, node, join_by = "|")
}
\description{
Helper to expand nodal expression
}
