% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tracing_strategies.R
\name{expected_learning}
\alias{expected_learning}
\title{Expected learning}
\usage{
expected_learning(model, query, strategy, given = NULL, lambda = NULL)
}
\arguments{
\item{model}{A  model}

\item{query}{A query as a character string, for example 'Y[X=1]>Y[X=0]'}

\item{strategy}{A set of variables to be sought}

\item{given}{A conditioning set as a character string that evaluates to a logical, for example 'Y==1'}

\item{lambda}{a parameter vector}
}
\description{
Expected reduction in variance from one step data collection strategy
}
\examples{
# Reduction in variance given monotonic X -> M -> Y model
library(dplyr)
model <- make_model("X" \%->\% "M1", "M1" \%->\% "M2", "M2" \%->\% "Y") \%>\%
  set_restrictions(node_restrict = list(M1 = "10", Y = "10")) \%>\%
  set_lambda(average = TRUE)
el <- expected_learning(model, model$lambda, query = "Y[X=1]>Y[X=0]",
                  strategy = c("M1", "M2"), given = "Y==1")
attr(el, "results_table")
}
