% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_strategies.R
\name{fill_bucket}
\alias{fill_bucket}
\title{helper to fill buckets dataframe}
\usage{
fill_bucket(model, buckets, vars, row = 1, column = 4)
}
\arguments{
\item{buckets}{dataframe with columns event, count and capacity vars plus strategy allocation var}

\item{vars}{vars to be observed}
}
\description{
helper to fill buckets dataframe
}
\examples{
model <- make_model("X->M->Y")
buckets = data.frame(event = "X0Y0", count = 3, capacity = 3, strategy = 2)
# Find different data that might result from looking at "M" in 2 out of 3 X0Y0 data types
fill_bucket(model, buckets, vars = "M")
}
