% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{find_rounding_threshold}
\alias{find_rounding_threshold}
\title{helper to find rounding thresholds for print methods}
\usage{
find_rounding_threshold(x)
}
\arguments{
\item{x}{An object for rounding}
}
\description{
helper to find rounding thresholds for print methods
}
