% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbiqq-package.R
\docType{package}
\name{gbiqq-package}
\alias{gbiqq-package}
\title{gbiqq: A package for creating and updating  binary probabilistic causal models}
\description{
gbiqq: A package for creating and updating  binary probabilistic causal models
}
