% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan_using_dags.R
\name{gbiqq}
\alias{gbiqq}
\title{Fit stan model}
\usage{
gbiqq(dag, lambda_priors, data, ...)
}
\arguments{
\item{dag}{A dag as created by \code{make_dag}}

\item{data}{A data frame with observations}

\item{lambdas_prior}{A vector containg priors for lambda}
}
\description{
Fit stan model
}
