% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbiqq_designer.R
\name{gbiqq_designer}
\alias{gbiqq_designer}
\title{Integrate gbiqq with DeclareDesign}
\usage{
gbiqq_designer(model = make_model("X -> Y"), restrictions = NULL,
  priors = "uniform", parameters = NULL, data_strat = NULL,
  queries = list(ATE = "Y[X=1] - Y[X=0]"))
}
\arguments{
\item{model}{A model created by make_model()}

\item{data_strat}{should be of the form "list(n_obs = n_obs, vars = list("X", "Y"), probs = list(...), n = NULL, subsets = list(...))}
}
\description{
Define a model, estimand, and answer strategy
}
\examples{
my_design <- gbiqq_designer(
  model = make_model("X -> Y"),
  parameters = c(.5, .5, .1, .7, .1, .1),  # Truth
  data_strat = list(n_obs = 5, vars = list(NULL), probs = list(NULL), ns = NULL,
     subsets = list(NULL)),
  queries = list(ATE = "Y[X=1] - Y[X=0]"))
DeclareDesign:::draw_data(my_design)
DeclareDesign:::draw_estimands(my_design)
DeclareDesign:::draw_estimates(my_design)
# DeclareDesign:::diagnose_design(my_design, sims = 2)
}
