% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_inherit_params.R
\name{gbiqq_internal_inherit_params}
\alias{gbiqq_internal_inherit_params}
\title{Create parameter documentation to inherit}
\usage{
gbiqq_internal_inherit_params(
  model,
  query,
  join_by,
  parameters,
  P,
  A,
  data,
  data_events,
  node,
  statement,
  using,
  n_draws
)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}

\item{query}{A character. An expression defining nodal types to interrogate \code{\link{reveal_outcomes}}}

\item{join_by}{A character. The logical expression connecting expanded types, \emph{AND} ('&') or \emph{OR} ('|'). Defaults to '|'.}

\item{parameters}{A vector of real numbers in [0,1]. Values of parameters may be specified. By default, parameters is drawn from model$parameters_df.}

\item{P}{A \code{data.frame}. Parameter matrix. Not required but may be provided to avoid repeated computation for simulations.}

\item{A}{A \code{data.frame}. Ambiguity matrix. Not required but may be provided to avoid repeated computation for simulations.}

\item{data}{A \code{data.frame}.  Data of nodes that can take three values: 0, 1, and NA. In long form as made by \code{\link{simulate_events}}}

\item{data_events}{A \code{data.frame}. It must be compatible with nodes in \code{model}. The default columns are \code{event}, \code{strategy} and \code{count}.}

\item{node}{A character. The quoted name of a node.}

\item{statement}{A character. A quoted causal statement.}

\item{using}{A character. It indicates whether to use `priors`, `posteriors` or `parameters`.}

\item{n_draws}{An integer. If no prior distribution is provided, generate prior distribution with \code{n_draws} number of draws.}
}
\description{
Create parameter documentation to inherit
}
\keyword{internal}
