% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get}
\alias{get}
\title{Get}
\usage{
get(model, object = "dag")
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by
\code{\link{make_model}}.}

\item{object}{A character string specifying the component to retrieve.
Available options are:
\describe{
  \item{\code{"nodal_types"}}{}}}
}
\value{
Elements from a \code{causal_model} as specified
}
\description{
Returns specified elements from a \code{causal_model}.
Users can use \code{get} to retrieve various components including
nodal types, causal types, parameter priors, parameter posteriors,
type priors, type posteriors, and other relevant elements.
See argument \code{object} for other
}
