% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_deprecate.R
\name{get_ambiguities}
\alias{get_ambiguities}
\title{Higher level function that returns all of the type ambiguities for all possible patterns of evidence}
\usage{
get_ambiguities(model)
}
\arguments{
\item{model}{A dag created by make_dag()}
}
\value{
A list of mappings between types
}
\description{
Higher level function that returns all of the type ambiguities for all possible patterns of evidence
}
