% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_ambiguity_matrices.R
\name{get_ambiguities_matrix}
\alias{get_ambiguities_matrix}
\title{Get ambiguities matrix}
\usage{
get_ambiguities_matrix(dag)
}
\arguments{
\item{dag}{A dag as created by \code{make_dag}}
}
\value{
A data frame that maps types (columns) to possible data realizations (rows).
}
\description{
Get matrix that maps types to data realizations
}
\examples{

XMYdag <- make_dag(add_edges(parent = "X", children = c("M")),
                   add_edges(parent = "M", children = c("Y")))

get_ambiguity_matrix(dag = XMYdag)

}
