% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_ambiguity_matrices.R
\name{get_ambiguity_matrix}
\alias{get_ambiguity_matrix}
\title{FUNCTION_TITLE}
\usage{
get_ambiguity_matrix(dag)
}
\arguments{
\item{dag}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
