% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_deprecate.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Get list of ancestor variables in a dag}
\usage{
get_ancestors(model)
}
\arguments{
\item{A}{probabilistic causal model created by make_model()}
}
\value{
A list of ancestors in a DAG
}
\description{
Get list of ancestor variables in a dag
}
\examples{

\dontrun{
dag <- make_dag(
 add_edges(parent = "X",children = c("K","Y")),
 add_edges(parent = "K",children = "Y")
)

get_ancestors(dag)
}

}
