% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_args_for}
\alias{get_args_for}
\title{helper to extract arguments for a specific function}
\usage{
get_args_for(fun, dots)
}
\arguments{
\item{fun}{a function to be checked}

\item{dots}{a named list with possible arguments}
}
\description{
helper to extract arguments for a specific function
}
