% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_models.R
\name{get_causal_types}
\alias{get_causal_types}
\title{Get expanded types}
\usage{
get_causal_types(model)
}
\arguments{
\item{model}{A model as created by \code{make_model}}
}
\value{
types
}
\description{
Create a data frame with types produced from all combinations of possible data produce by a dag.
}
