% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_dags.R
\name{get_chains}
\alias{get_chains}
\title{Create list by endogenous variable containing all chains terminating at this endogenous variable}
\usage{
get_chains(pcm)
}
\arguments{
\item{dag}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
test_dag <-
gbiqq::make_dag(add_edges(parent = "X",children = c("K", "Y1")),
								add_edges(parent = "K", children = c("Y1","Y2")),
								add_edges(parent = "Z", children = c("K")))

get_chains(test_dag)
}

}
