% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_dags.R
\name{get_children_of_exogenous}
\alias{get_children_of_exogenous}
\title{Get children of exogenous variables}
\usage{
get_children_of_exogenous(dag)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
A list of the children of exogenous variables
}
\description{
Get children of exogenous variables
}
