% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_confounds_df.R
\name{get_confounds_df}
\alias{get_confounds_df}
\title{Get confounds dataframe}
\usage{
get_confounds_df(model, confounds_df = NULL, ...)
}
\arguments{
\item{model}{A model made by make_model}

\item{confounds_df}{A dataframe of pairs of nodes that have unobserved confounding}

\item{...}{arguments passed to make_confounds_df}
}
\description{
Gets matrix of confounded nodal types from model.
}
\examples{
model <- make_model("X -> M -> Y; X <-> M;  M <-> Y") \%>\%
   set_confounds_df()
get_confounds_df(model)


}
