% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{get_data_families}
\alias{get_data_families}
\title{get_data_families}
\usage{
get_data_families(
  model,
  drop_impossible = TRUE,
  drop_none = TRUE,
  mapping_only = FALSE
)
}
\arguments{
\item{model}{A probabilistic causal model created by make_model().}

\item{drop_impossible}{Logical. Drops data that is impossible given model restrictions.}

\item{drop_none}{Logical. Drops row for no data.}

\item{mapping_only}{Logical. Returns data mapping matrix only.}
}
\value{
Returns indices and ambiguity matrix
}
\description{
Get possible data types
}
\examples{
get_data_families(model = make_model("X->Y"))
get_data_families(model = make_model("X->Y"), mapping_only = TRUE)
get_data_families(model = make_model("X-> M -> Y"))


}
