% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_dags.R
\name{get_endogenous_vars}
\alias{get_endogenous_vars}
\title{Get endogenous variables in a DAG}
\usage{
get_endogenous_vars(pcm)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
A vector of endogenous variables
}
\description{
Get endogenous variables in a DAG
}
