% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_prob.R
\name{get_event_prob}
\alias{get_event_prob}
\title{Draw event probabilities}
\usage{
get_event_prob(model, P = NULL, A = NULL, parameters = NULL, type_prob = NULL)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{P}{A matrix. Parameter matrix. Not required but may be provided to avoid repeated computation for simulations.}

\item{A}{A matrix. Ambiguity matrix. Not required but may be provided to avoid repeated computation for simulations.}

\item{parameters}{A numeric vector. Values of parameters may be specified. By default, parameters is drawn from model$parameters_df.}

\item{type_prob}{A numeric vector. Type probabilities (usually not required).}
}
\description{
Draw event probabilities
}
\examples{
model <- make_model('X -> Y')
get_event_prob(model = model)
get_event_prob(model = model, parameters = rep(1, 6))
get_event_prob(model = model, parameters = 1:6)
}
