% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_models.R
\name{get_exogenous_vars}
\alias{get_exogenous_vars}
\title{Get exogenous variables in a DAG}
\usage{
get_exogenous_vars(model)
}
\arguments{
\item{A}{probabilistic causal model created by make_model()}
}
\value{
A vector of exogenous variables
}
\description{
Get exogenous variables in a DAG
}
