% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguities.R
\name{get_indicator_matrix}
\alias{get_indicator_matrix}
\title{Get parameter matrix}
\usage{
get_indicator_matrix(model)
}
\arguments{
\item{model}{A model created by make_model()}
}
\description{
Generate a  matricex that maps from parameters into causal types. In models without confounding parameters correspond to nodal types.
}
