% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_inital_parameter_names}
\alias{get_inital_parameter_names}
\title{get_initial_parameter_names if no confounding}
\usage{
get_inital_parameter_names(model)
}
\arguments{
\item{model}{A model}
}
\description{
get_initial_parameter_names if no confounding
}
