% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan.R
\name{get_max_possible_data}
\alias{get_max_possible_data}
\title{Get data frame with all possible data combinations}
\usage{
get_max_possible_data(model)
}
\arguments{
\item{A}{probabilistic causal model created by make_model()}
}
\value{
A data frame
}
\description{
Get data frame with all possible data combinations
}
