% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_deprecate.R
\name{get_n_endogenous_types}
\alias{get_n_endogenous_types}
\title{Get number of endogenous types}
\usage{
get_n_endogenous_types(model)
}
\arguments{
\item{A}{probabilistic causal model created by make_model()}
}
\value{
A list of the numbers of endogenous types
}
\description{
Get number of endogenous types
}
