% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_confounds_df.R
\name{get_nodal_joint_probability}
\alias{get_nodal_joint_probability}
\title{Calculate joint distribution of nodal types}
\usage{
get_nodal_joint_probability(model, parameters = NULL,
  generic_parameters = NULL)
}
\arguments{
\item{model}{A model made by make_model}

\item{parameters}{A parameter vector}

\item{generic_parameters}{Logical, require selection of random parameter.}
}
\description{
Identifies confounded nodal types. Default behavior draws a random admissible parameter vector.
}
\examples{
model <- make_model("X -> Y") \%>\%
  set_confound(list("X <-> Y"))

get_nodal_joint_probability(model)

# Default parameter values may not reveal dependencies
get_nodal_joint_probability(model, generic_parameters = FALSE)

}
