% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_ambiguity_matrices.R
\name{get_nodal_types}
\alias{get_nodal_types}
\title{Get nodal types
Nodal types are created by concatening variables and their possible data. Used for labeling ambiguities matrix.}
\usage{
get_nodal_types(pcm)
}
\arguments{
\item{pcm}{A dag as created by \code{make_dag}}
}
\value{
types
}
\description{
Get nodal types
Nodal types are created by concatening variables and their possible data. Used for labeling ambiguities matrix.
}
\keyword{internal}
