% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodal_types.R
\name{get_nodal_types_dep}
\alias{get_nodal_types_dep}
\title{Get nodal types}
\usage{
get_nodal_types_dep(model)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}
}
\value{
Nodal types
}
\description{
Nodal types are created by concatening variables and their possible data. Used for labeling ambiguities matrix.
}
