% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_dags.R
\name{get_observable_data}
\alias{get_observable_data}
\title{Get observable data implied by DAG}
\usage{
get_observable_data(variables)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
A list of data realizations
}
\description{
Get observable data implied by DAG
}
