% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_parameter_names}
\alias{get_parameter_names}
\title{get_parameter_names}
\usage{
get_parameter_names(model, include_paramset = TRUE)
}
\arguments{
\item{model}{A model}

\item{include_paramset}{Logical, whether to include the param set prefix as part of the name.}
}
\description{
Parameter names taken from P matrix or model if no P matrix provided
}
\examples{

get_parameter_names(make_model("X->Y"))

}
