% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_dags.R
\name{get_parents}
\alias{get_parents}
\title{Get list of parent variables in a dag}
\usage{
get_parents(dag)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
A list of parents in a DAG
}
\description{
Get list of parent variables in a dag
}
\examples{

dag <- make_dag(
 add_edges(parent = "X",children = c("K","Y")),
 add_edges(parent = "K",children = "Y")
)

get_parents(dag)


}
