% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmap.R
\name{get_parmap}
\alias{get_parmap}
\title{Get parmap: a matrix mapping from parameters to data types}
\usage{
get_parmap(model)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}
}
\value{
A matrix
}
\description{
Gets parmap from a model, or generates if not available.
}
\examples{
get_parmap(model = make_model('X->Y'))

}
