% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan_using_dags.R
\name{get_pi_expanders}
\alias{get_pi_expanders}
\title{Get helpers for expanding the pi matrix}
\usage{
get_pi_expanders(pi, pcm)
}
\arguments{
\item{pcm}{A dag created by make_dag()}
}
\value{
A list of pi helpers
}
\description{
Get helpers for expanding the pi matrix
}
