% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan_using_dags.R
\name{get_possible_data}
\alias{get_possible_data}
\title{This creates a list of all possible data realizations, for each child}
\usage{
get_possible_data(dag)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
A list of all data realizations possible
}
\description{
This creates a list of all possible data realizations, for each child
}
