% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan_using_dags.R
\name{get_possible_data_internal}
\alias{get_possible_data_internal}
\title{get_possible_data_internal}
\usage{
get_possible_data_internal(pcm, collapse = TRUE)
}
\arguments{
\item{pcm}{A dag created by \code{make_dag}}
}
\value{
a list
}
\description{
get_possible_data_internal
}
\keyword{internal}
