% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_distribution.R
\name{get_posterior_distribution}
\alias{get_posterior_distribution}
\title{Get the posterior distribution from a model}
\usage{
get_posterior_distribution(model)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by
\code{\link{make_model}}.}
}
\value{
A `data.frame` with parameters draws
}
\description{
Access the posterior distribution from the model if one has been added via `update_model`.
}
\concept{posterior_distribution}
\keyword{internal}
