% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_priors}
\alias{get_priors}
\title{Get prior distribution}
\usage{
get_priors(model, priors = NULL, prior_distribution = "uniform",
  alphas = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{priors}{Vector of Dirichlet hyperparameters}

\item{prior_distribution}{A character indicating the prior distribuiton. One of "jeffreys", "certainty", "uniform".}

\item{alphas}{the hyperparameters of the dirichlet distribution.}
}
\description{
Get prior distribution
}
