% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_and_estimands.R
\name{get_single_estimand}
\alias{get_single_estimand}
\title{Calculate estimand}
\usage{
get_single_estimand(model, query, subset = TRUE, lambda = NULL,
  posterior = FALSE, type_distribution = NULL, verbose = TRUE)
}
\arguments{
\item{model}{A  model}

\item{query}{A query on potential otcomes such as "Y[X=1] - Y[X=0]"}

\item{subset}{quoted expression evaluates to logical statement. subset allows estimand to be conditioned on *observational* distribution.}

\item{lambda}{A  true parameter vector, if available}

\item{posterior}{if true use a posterior distribution, otherwise use the prior}

\item{type_distribution}{if provided saves calculation, otherwise clculated from model; may be based on prior or posterior}
}
\description{
Calculated from a prior or posterior distribution
}
\examples{
model <- make_model("X" \%->\% "Y") \%>\%
         set_prior_distribution()
 estimand_1  <- get_single_estimand(model, query = "Y[X=1] - Y[X=0]")
 estimand_2  <- get_single_estimand(model, query = "Y[X=1] > Y[X=0]")
 get_single_estimand(model, lambda = draw_lambda(model), query = "Y[X=1] > Y[X=0]")
}
