% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_restrictions.R
\name{get_type_names}
\alias{get_type_names}
\title{Get type names}
\usage{
get_type_names(nodal_types)
}
\arguments{
\item{nodal_types}{Nodal types of a model. See \code{get_nodal_types()}.}
}
\description{
Get type names
}
