% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_model.R
\name{get_type_posteriors}
\alias{get_type_posteriors}
\title{helper to get type distributions}
\usage{
get_type_posteriors(jobs, model, n_draws, parameters = NULL)
}
\arguments{
\item{jobs}{data frame of argument combinations}

\item{model}{a list of models}

\item{n_draws}{integer specifying number of draws from prior distribution}

\item{parameters}{optional list of parameter vectors}
}
\value{
jobs data frame with a nested column of type distributions
}
\description{
helper to get type distributions
}
\keyword{internal}
