% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_type_prob_multiple_c}
\alias{get_type_prob_multiple_c}
\title{generates n draws from type probability distribution for each type in P}
\usage{
get_type_prob_multiple_c(params, P)
}
\arguments{
\item{params}{parameters, priors or posteriors}

\item{P}{parameter_matrix of parameters and causal types}
}
\value{
draws from type distribution for each type in P
}
\description{
generates n draws from type probability distribution for each type in P
}
