% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_models.R
\name{get_variables}
\alias{get_variables}
\title{Get variable names from a DAG}
\usage{
get_variables(model)
}
\arguments{
\item{A}{probabilistic causal model created by make_model()}
}
\value{
A vector of variable names
}
\description{
Get variable names from a DAG
}
