% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{get_weights_ambiguities}
\alias{get_weights_ambiguities}
\title{get_weights_ambiguities}
\usage{
get_weights_ambiguities(model, data)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
Returns indices and ambiguity matrix
}
\description{
get_weights_ambiguities
}
