% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gsub_many}
\alias{gsub_many}
\title{Recursive substitution}
\usage{
gsub_many(x, pattern_vector, replacement_vector, ...)
}
\arguments{
\item{x}{A character vector.}

\item{pattern_vector}{A character vector.}

\item{replacement_vector}{A character vector.}

\item{...}{Options passed onto \code{gsub()} call.}
}
\value{
Returns multiple expression with substituted elements
}
\description{
Applies \code{gsub()} from multiple patterns to multiple
replacements with 1:1 mapping.
}
\keyword{internal}
