% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{institutions_data}
\alias{institutions_data}
\title{Institutions and growth: Data for replication of analysis in
*Integrated Inferences*}
\format{
A data frame with 79 rows and 5 columns:
\describe{
  \item{Y}{Income (GDP PPP 1995), dichotomized}
  \item{R}{Institutions, (based on  Kaufmann, Kraay, and Zoido-Lobaton (2002)) dichotomized}
  \item{D}{Distance from the equator (in degrees), dichotomized}
  \item{M}{Settler mortality (from Acemoglu, Johnson, and Robinson), dichotomized}
  \item{country}{Country}
}
}
\source{
\url{https://drodrik.scholar.harvard.edu/publications/institutions-rule-primacy-institutions-over-geography-and-integration}
}
\usage{
institutions_data
}
\description{
A  dataset containing dichotomized versions of variables in
Rodrik, Subramanian, and Trebbi (2004).
}
\keyword{datasets}
