% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{interacted_nodes}
\alias{interacted_nodes}
\title{function to generate nodal types for interactions}
\usage{
interacted_nodes(
  nodal_types,
  nodes = NULL,
  interaction_order = 2L,
  interaction_operation = c("&", "|")
)
}
\arguments{
\item{nodal_types}{a named list of nodal types}

\item{nodes}{a character vector specifying the names of nodes to generate interactions for. Defaults to all nodes}

\item{interaction_order}{an integer specifying the type of interaction. Defaults to two-way interaction}

\item{interaction_operation}{the logical operations to apply to construct interactions. Defaults to AND and OR operation for standard interactions.}
}
\value{
a named list of previous and interaction nodal types for all nodes
}
\description{
function to generate nodal types for interactions
}
