% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_a_model}
\alias{is_a_model}
\title{Check whether argument is a model}
\usage{
is_a_model(model)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by
\code{\link{make_model}}.}
}
\value{
An error message if argument is not a model.
}
\description{
Check whether argument is a model
}
\keyword{internal}
