% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_improper}
\alias{is_improper}
\title{Check whether a model is improper}
\usage{
is_improper(model)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by
\code{\link{make_model}}.}
}
\value{
A logical expression indicating whether a model is improper
}
\description{
Compute the sum of causal types probabilities. A model is flagged as
  improper if the sum is different than one.
}
\keyword{internal}
