% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{layout_dag_coords}
\alias{layout_dag_coords}
\title{Sugiyama coordinates from directed edges; include confound-only nodes.}
\usage{
layout_dag_coords(dag)
}
\description{
Sugiyama coordinates from directed edges; include confound-only nodes.
}
\keyword{internal}
