% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipids_data}
\alias{lipids_data}
\title{Lipids: Data for Chickering and Pearl replication}
\format{
A data frame with 8 rows and 3 columns:
\describe{
  \item{event}{The data type}
  \item{strategy}{For which nodes is data available}
  \item{count}{Number of units with this data type}
}
}
\source{
\url{https://cdn.aaai.org/AAAI/1996/AAAI96-188.pdf}
}
\usage{
lipids_data
}
\description{
A compact dataset containing information on an encouragement, (Z, cholestyramine prescription), a treatment (X, usage), and an outcome (Y, cholesterol).
From David Maxwell Chickering and Judea Pearl: "A Clinician’s Tool for Analyzing Non-compliance", AAAI-96 Proceedings. Chickering and Pearl in turn draw the data from Efron, Bradley, and David Feldman. "Compliance as an explanatory variable in clinical trials." Journal of the American Statistical Association 86.413 (1991): 9-17.
}
\keyword{datasets}
