% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type_helpers.R
\name{list_non_parents}
\alias{list_non_parents}
\title{Returns a list with the nodes that are not directly pointing into a node}
\usage{
list_non_parents(model, node)
}
\arguments{
\item{node}{A quoted name of a node}

\item{statement}{A quoted causal statement.}
}
\description{
Returns a list with the nodes that are not directly pointing into a node
}
\keyword{internal}
