% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_type.R
\name{lookup_nodal_type_internal}
\alias{lookup_nodal_type_internal}
\title{Reveal nodal types according to a query}
\usage{
lookup_nodal_type_internal(model, query)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}

\item{query}{A character string. An expression defining nodal types to interrogate \code{\link{reveal_outcomes}}}
}
\description{
Reveal nodal types according to a query
}
