% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_type.R
\name{lookup_type_internal}
\alias{lookup_type_internal}
\title{Reveal nodal types according to a query}
\usage{
lookup_type_internal(model, query)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{query}{A character vector of length 1L. An expression in string format defining causal types to interrogate \code{reveal_outcomes()}}
}
\description{
Reveal nodal types according to a query
}
