% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguities.R
\name{make_adjacency_matrix}
\alias{make_adjacency_matrix}
\title{FUNCTION_TITLE}
\usage{
make_adjacency_matrix(model)
}
\arguments{
\item{model}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
