% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_alphas}
\alias{make_alphas}
\title{Make alphas}
\usage{
make_alphas(model, alphas = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{alphas}{the hyperparameters of the dirichlet distribution.
Stipulated alpha values override prior_distribution}
}
\description{
Make alphas
}
\examples{
XYmodel <- make_model("X -> Y")
#  Default sets all priors to 1
make_priors(model = XYmodel)
#  Set all priors to 0.5
make_priors(model = XYmodel, prior_distribution = "jeffreys")
#  Set all priors to infinity
make_priors(model = XYmodel, prior_distribution = "certainty")

#  set all priors to 1 except for prior of nodal_type X0
make_priors(model = XYmodel, alphas = list(X = c(X0 = 2)))
#  specify priors for each of the nodal_types in model
make_priors(model = XYmodel,
           alphas = list(X = c(X0 = 2, `X == 1` = 3),  Y = c(Y00 = 1, `(Y[X=1] > Y[X=0])` = 3, Y01 = 2, `(Y[X=1] == Y[X=0])`  = 3)))
# set all priors to 10
make_priors(model = XYmodel, alphas =  10)


}
