% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ambiguities_matrix.R
\name{make_ambiguities_matrix}
\alias{make_ambiguities_matrix}
\title{Make ambiguities matrix}
\usage{
make_ambiguities_matrix(model)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}
}
\value{
A data frame. Types (rows) corresponding to possible data realizations (columns).

A matrix representing ambiguities.
}
\description{
Make ambiguities matrix. The ambiguities matrix  maps from causal types into data types.
}
\examples{
model <- make_model('X -> Y')
make_ambiguities_matrix(model = model)

}
