% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan_using_dags.R
\name{make_ambiguity_matrices}
\alias{make_ambiguity_matrices}
\title{Higher level function that returns an indicator matrix showing correspondence between type and possible evidence pattern for each type and evidence pattern}
\usage{
make_ambiguity_matrices(pcm)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
A list of mappings between types
}
\description{
Higher level function that returns an indicator matrix showing correspondence between type and possible evidence pattern for each type and evidence pattern
}
