% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan_using_dags.R
\name{make_ambiguity_matrix_internal}
\alias{make_ambiguity_matrix_internal}
\title{Internal function that transforms the ambiguity list into an indicator
function}
\usage{
make_ambiguity_matrix_internal(ambiguities)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\description{
Internal function that transforms the ambiguity list into an indicator
function
}
