% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Make data}
\usage{
make_data(
  model,
  n = 1,
  parameters = NULL,
  param_type = NULL,
  vars = NULL,
  n_steps = NULL,
  probs = NULL,
  subsets = TRUE,
  complete_data = NULL,
  ...
)
}
\arguments{
\item{model}{A causal model as created by \code{make_model}}

\item{n}{Scalar giving number of observations in \code{complete_data}}

\item{parameters}{A specific parameter vector. If not provided parameters are drawn using model$parameter_df}

\item{param_type}{A character string specifying type of parameters to make ("flat", "prior_mean", "posterior_mean", "prior_draw", "posterior_draw", "define). With param_type set to \code{define} use arguments to be passed to \code{make_priors}; otherwise \code{flat} sets equal probabilities on each nodal type in each parameter set; \code{prior_mean}, \code{prior_draw}, \code{posterior_mean}, \code{posterior_draw} take parameters as the means or as draws from the prior or posterior.}

\item{vars}{List indicating number which nodes to be observed at each step}

\item{n_steps}{List indicating number of observations to be observed at each step}

\item{probs}{List indicating observation probabilities at each step}

\item{subsets}{List indicating strata within which observations are to be observed at each step}

\item{complete_data}{A dataset with complete observations. Optional.}

\item{...}{additional arguments that can be passed to make_parameters}
}
\value{
A data frame of simulated data.
}
\description{
Make data
}
\examples{


# Simple draws
model <- make_model("X -> M -> Y")
make_data(model)
make_data(model, n = 3, vars = c("X","Y"))
make_data(model, n = 3, param_type = "prior_draw")
make_data(model, n = 10, param_type = "define", alpha =  0:9)

# Data Strategies
# A strategy in which X, Y are observed for sure and M is observed
# with 50\% probability for X=1, Y=0 cases

model <- make_model("X -> M -> Y")
make_data(
  model,
  n = 8,
  vars = list(c("X", "Y"), "M"),
  probs = list(1, .5),
  subsets = list(NULL, "X==1 & Y==0"))

# Simulate multiple datasets is fastest if w is provided
model <- make_model("X -> Y")
w <- get_event_prob(model)
replicate(5, make_data_single(model, n = 5, w = w))

}
