% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_strategies.R
\name{make_data_probabilities}
\alias{make_data_probabilities}
\title{Generates a probability distribution over possible data outcomes}
\usage{
make_data_probabilities(model, given, data_strat = NULL, pars)
}
\arguments{
\item{model}{A causal model as created by \code{make_model}}

\item{given}{A data frame with observations}

\item{data_strat}{data strategy}
}
\value{
A dataset
}
\description{
THIS IS NOT GENERAL AT ALL YET. THE BASIC IDEA THOUGH IS TO USE draw_event_prob
THOUGH IN PRACTICE WE WILL LIKELY NEED dmultinom
}
\examples{

library(dplyr)
model <- make_model("X->M->Y")  \%>\%
   set_restrictions(causal_type_restrict = "Y[M=1]<Y[M=0] | M[X=1]<M[X=0] ") \%>\%
   set_parameter_matrix()

given = data.frame(X = c(0,0,0,1,1,1), M = NA, Y = c(0,0,1,0,1,1))
pars <- draw_parameters(model)
make_data_probabilities(model, given, data_strat = c(1,3), pars = pars)

}
