% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_strategies.R
\name{make_estimates_database}
\alias{make_estimates_database}
\title{Generates a database of results using gbiqq over possible data}
\usage{
make_estimates_database(model, given, possible_data = NULL,
  queries = "Y[X=1]>Y[X=0]", data_strat = NULL)
}
\arguments{
\item{model}{A causal model as created by \code{make_model}}

\item{possible_data}{A data frame with possible data}

\item{queries}{Queries}
}
\value{
A dataframe
}
\description{
This function runs many models and can take a long time depending on the size of possible data.
}
\examples{
#' library(dplyr)
model <- make_model("X->M->Y")  \%>\%
   set_restrictions(causal_type_restrict = "Y[M=1]<Y[M=0] | M[X=1]<M[X=0] ") \%>\%
   set_parameter_matrix()

given = data.frame(X = c(0,0,0,1,1,1), M = NA, Y = c(0,0,1,0,1,1))

possible_data <- make_possible_data(model, given)

estimates_database <- make_estimates_database(
      model,
      given = given,
      possible_data,
      queries = "Y[X=1]>Y[X=0]")

}
