% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodal_types.R
\name{make_nodal_types}
\alias{make_nodal_types}
\title{make nodal types}
\usage{
make_nodal_types(model, include_node_names = FALSE)
}
\arguments{
\item{model}{A causal model}
}
\description{
make nodal types
}
