% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_matrix.R
\name{make_parameter_matrix}
\alias{make_parameter_matrix}
\title{Make parameter matrix}
\usage{
make_parameter_matrix(model)
}
\arguments{
\item{model}{A model created by make_model()}

\item{confound}{A list relating nodes to types with whihc  they are confounded}
}
\description{
Calculate parameter matrix assuming no confounding. The parameter matrix  maps from parameters into causal types. In models without confounding parameters correspond to nodal types.
}
\examples{
model <- make_model("X -> Y")
get_parents(model)
confound = list(ancestor = c(X="X"), descendent_type = list( Y = c("00","11") ))
make_parameter_matrix(model = model, confound = confound)
model <- make_model("X -> M -> Y")
get_parents(model)
confound = list(ancestor = c(X="X"), descendent_type = list( Y = c("00","11") ))
make_parameter_matrix(model = model, confound = confound)
}
