% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_parameters.R
\name{make_parameters}
\alias{make_parameters}
\title{Make parameters
Default behavior takes a random draw of parameters from priors; can also draw from posteriors, or return pre-defined parameters.}
\usage{
make_parameters(model, using = NULL)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{using}{String indicating whether to use `priors`, `posteriors` or `parameters`}
}
\value{
A named list of parameters.
}
\description{
Make parameters
Default behavior takes a random draw of parameters from priors; can also draw from posteriors, or return pre-defined parameters.
}
\examples{
make_parameters(model = make_model("X->Y"))

}
