% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{make_parameters_df}
\alias{make_parameters_df}
\title{make_parameters_df(list(X = "1", Y = c("01", "10")))}
\usage{
make_parameters_df(nodal_types)
}
\description{
make_parameters_df(list(X = "1", Y = c("01", "10")))
}
