% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmap.R
\name{make_parmap}
\alias{make_parmap}
\title{Make parmap: a matrix mapping from parameters to data types}
\usage{
make_parmap(model, A = NULL, P = NULL)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by
\code{\link{make_model}}.}

\item{A}{A \code{data.frame}. Ambiguities matrix. Not required but may be
provided to avoid repeated computation for simulations. \code{inspect(model, "ambiguities_matrix")}}

\item{P}{A \code{data.frame}. Parameter matrix. Not required but may be
provided to avoid repeated computation for simulations. See \code{inspect(model, "parameter_matrix")}.}
}
\value{
A matrix
}
\description{
Generates a matrix with a row per parameter and a column per data type.
}
