% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_prior_distribution.R
\name{make_prior_distribution}
\alias{make_prior_distribution}
\title{Make a prior distribution from priors}
\usage{
make_prior_distribution(model, n_draws = 4000)
}
\arguments{
\item{model}{A model object generated by make_model().}

\item{n_draws}{A scalar. Number of draws.}
}
\description{
Add n_param x n_draws database of possible lambda draws to the model.
}
\examples{
make_model("X -> Y") \%>\% make_prior_distribution(n_draws = 5)

}
\seealso{
Other prior_distribution: 
\code{\link{get_prior_distribution}()},
\code{\link{set_prior_distribution}()}
}
\concept{prior_distribution}
