% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_priors}
\alias{make_priors}
\title{Make Priors}
\usage{
make_priors(model = list(model), prior_distribution = NA,
  alphas = NULL, node = NA, statement = NA, confound = NA,
  label = NA)
}
\arguments{
\item{model}{A mode created with \code{make_model}}

\item{prior_distribution}{A common prior distribution (uniform, jeffreys or certainty)}

\item{alphas}{Hyperparameters of the Dirichlet distribution}

\item{node}{A node}

\item{statement}{A query}

\item{confound}{A confound statement}
}
\description{
A flexible function to add priors to a model.
}
\details{
Four arguments govern *which* parameters should be altered. The default is "all" but this can be reduced by specifying
* \code{label} The label of a particular nodal type, written either in the form Y0000 or Y.Y000
* \code{node}, which restricts for example to parameters associated with node "X"
* \code{statement}, which restricts for example to nodal types that satisfy the statment "Y[X=1] > Y[X=0]"
* \code{confound}, which restricts for example to nodal types that satisfy the statment "Y[X=1] > Y[X=0]"

Two arguments govern what values to apply: alphas is one or more non negative numbers and "prior" distribution indicates one of a common class: uniform, jeffreys, or "certain"
}
\examples{
model <- make_model("X -> Y")
make_priors(model, node = list("X", "Y"), alphas = list(3, 6))
model <- make_model("X->Y")
model <- set_confound(model, list(X = "Y[X=1] > Y[X=0]"))
model <- set_confound(model, list(X = "Y[X=1] < Y[X=0]"))
make_priors(model, confound = list("Y[X=1] > Y[X=0]", "Y[X=1] < Y[X=0]"), alphas = list(3, 6))
}
