% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan_using_dags.R
\name{map_types_to_data}
\alias{map_types_to_data}
\title{Map types of every varibale to observable data}
\usage{
map_types_to_data(pcm)
}
\arguments{
\item{pcm}{A dag created by make_dag()}
}
\value{
A list of mappings between types and data
}
\description{
Map types of every varibale to observable data
}
