% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{n_check}
\alias{n_check}
\title{n_check}
\usage{
n_check(n)
}
\arguments{
\item{n}{An integer. Sample size argument.}
}
\value{
An error message if \code{n} is not an integer or is less than 0.
}
\description{
n_check
}
\details{
Checks whether the input is an integer greater than 0.
}
\keyword{internal}
