% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type_helpers.R
\name{nodes_in_statement}
\alias{nodes_in_statement}
\title{Identify nodes in statement}
\usage{
nodes_in_statement(nodes, statement)
}
\arguments{
\item{nodes}{A vector with quoted names of the nodes}

\item{statement}{A quoted causal statement.}
}
\description{
Identify nodes in statement
}
\keyword{internal}
