% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_helpers.R
\name{non_increasing}
\alias{non_increasing}
\title{Make monotonicity statement (non positive)}
\usage{
non_increasing(X, Y)
}
\arguments{
\item{X}{A character. The quoted name of the input node}

\item{Y}{A character. The quoted name of the outcome node}
}
\value{
A character statement of class \code{statement}
}
\description{
Generate a statement for Y weakly monotonic (not increasing) in X
}
\examples{
non_increasing('A', 'B')

}
\seealso{
Other statements: 
\code{\link{complements}()},
\code{\link{decreasing}()},
\code{\link{increasing}()},
\code{\link{interacts}()},
\code{\link{non_decreasing}()},
\code{\link{substitutes}()},
\code{\link{te}()}
}
\concept{statements}
