% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_strategies.R
\name{observe}
\alias{observe}
\title{Observe data, given a strategy}
\usage{
observe(complete_data, observed = NULL, vars_to_observe = NULL,
  prob = 1, m = NULL, subset = NULL)
}
\arguments{
\item{complete_data}{A data.frame. Data observed and unobserved.}

\item{observed}{A data.frame. Data observed.}

\item{vars_to_observe}{A list of variables to observe.}

\item{prob}{A scalar. Observation probability.}

\item{m}{Number of units to observe; if specified, \code{m} overrides \code{prob}.}

\item{subset}{A logical statement that can be applied to rows of complete data. For instance observation fo some variables might depend on observed values of other variables; or observation may only be sought if data not already observed!}
}
\description{
Observe data, given a strategy
}
\examples{
model <- make_model("X -> Y")
df <- simulate_data(model, n = 8)
# Observe X values only
observe(complete_data = df, vars_to_observe = "X")
# Observe half the Y values for cases with observed X = 1
observe(complete_data = df,
     observed = observe(complete_data = df, vars_to_observe = "X"),
     vars_to_observe = "Y", prob = .5,
     subset = "X==1")
}
