% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{old_get_likelihood_helpers}
\alias{old_get_likelihood_helpers}
\title{get_likelihood_helpers}
\usage{
old_get_likelihood_helpers(pcm)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
Returns indices and ambiguity matrix
}
\description{
Produce the indices and ambiguity matrix needed for the
}
