% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{orientation}
\alias{orientation}
\title{internal helper
calculates the orientation of the point triplet (p, q, r)
this is used to check intersection between line segments}
\usage{
orientation(p, q, r)
}
\description{
internal helper
calculates the orientation of the point triplet (p, q, r)
this is used to check intersection between line segments
}
\keyword{internal}
