% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_models.R
\name{perm}
\alias{perm}
\title{Produces the possible permutations of a set of variables}
\usage{
perm(v)
}
\arguments{
\item{v}{A vector of integers indicating the number of values each variable takes on. E.g., a binary variable is represented by 2.}
}
\value{
A matrix of permutations
}
\description{
Produces the possible permutations of a set of variables
}
\examples{

\dontrun{
perm(c(2,2,2))
}
}
