% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_model.R
\name{plot.model_query}
\alias{plot.model_query}
\title{Plot model query results}
\usage{
\method{plot}{model_query}(x, ...)
}
\arguments{
\item{x}{An object of class \code{model_query}, usually from
\code{\link{query_model}}.}

\item{...}{Further arguments (currently unused; included for S3
compatibility).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot method for class \code{model_query}. Draws point estimates (and
credible intervals when present) from \code{\link{query_model}} output,
faceted by model when more than one model is in the table.
}
\examples{
\donttest{
model <- make_model("X -> Y")
q <- query_model(
  model,
  query = "Y[X=1] - Y[X=0]",
  using = "parameters"
)
plot(q)
}
}
