% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{plot_dag}
\alias{plot_dag}
\title{Plot your dag using dagitty}
\usage{
plot_dag(model, include_confounds = TRUE)
}
\arguments{
\item{model}{A model created by make_model()}

\item{include_confounds}{Logical. Adds confounds by default. If model does not contain a confounds_df this is generated on the fly.}
}
\value{
dagitty translation of DAG
}
\description{
Plot your dag using dagitty
}
\examples{
\dontrun{
model <- make_model("X -> K -> Y; X -> Y")
plot(model)

model <- make_model("X -> K -> Y; X <-> Y")
plot(model, include_confounds = FALSE)
plot(model)
plot_dag(model, include_confounds = TRUE)
}

}
