% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{position_nodes}
\alias{position_nodes}
\title{internal helper
position_nodes generates a force directed graph layout to optimally
position nodes. Connected nodes attract while unconnected nodes repel.
position_nodes additionally ensures that connected nodes are not placed
vertically or horizontally of each other i.e. that the slope of the edge
connecting them is not 0 or inf.}
\usage{
position_nodes(
  dag,
  nodes,
  iterations = 1000L,
  k = 0.5,
  repulsion = 0.1,
  desired_distance = 1
)
}
\description{
internal helper
position_nodes generates a force directed graph layout to optimally
position nodes. Connected nodes attract while unconnected nodes repel.
position_nodes additionally ensures that connected nodes are not placed
vertically or horizontally of each other i.e. that the slope of the edge
connecting them is not 0 or inf.
}
\keyword{internal}
