% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_estimation.R
\name{posterior_type_prob}
\alias{posterior_type_prob}
\title{Draw type probabilities after estimation}
\usage{
posterior_type_prob(updated_model)
}
\arguments{
\item{updated_model}{An updated model}
}
\description{
Draw type probabilities after estimation
}
