% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gbiqq_data.R
\name{prep_gbiqq_data}
\alias{prep_gbiqq_data}
\title{Prepare data for stan}
\usage{
prep_gbiqq_data(model, data)
}
\arguments{
\item{model}{A model created by \code{make_model}}

\item{data}{A  'compact' data frame (as made by `simulate_events()``)}
}
\value{
a list
}
\description{
Create a list containing the data to be passed to stan
}
\examples{
model <- make_model('X->Y')
data  <-  collapse_data(simulate_data(model, n = 6), model)
prep_gbiqq_data(model, data)

model <- make_model('X->Y') \%>\% set_confound(list(X = 'Y[X=1]>Y[X=0]'))
data  <-  collapse_data(simulate_data(model, n = 6), model)
prep_gbiqq_data(model, data)

}
