% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_causal_model.R
\name{print.causal_model}
\alias{print.causal_model}
\title{Print a short summary for a causal model}
\usage{
\method{print}{causal_model}(x, ...)
}
\arguments{
\item{x}{An object of \code{causal_model} class, usually a result of
a call to \code{\link{make_model}} or \code{\link{update_model}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{causal_model}.
}
\details{
The information regarding the causal model includes the statement describing
causal relations using arrow syntax (e.g. \code{"X -> Y"}),
number of nodal types per parent in a DAG, and number of causal types.
}
\keyword{internal}
