% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.causal_types}
\alias{print.causal_types}
\title{Print a short summary for causal_model causal-types}
\usage{
\method{print}{causal_types}(x, ...)
}
\arguments{
\item{x}{An object of \code{causal_types} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{make_model} or \code{update_model}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{causal_types}.
}
