% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{print.event_probabilites}
\alias{print.event_probabilites}
\title{Print a short summary for event probabilities}
\usage{
\method{print}{event_probabilites}(x, ...)
}
\arguments{
\item{x}{An object of \code{event_probabilities}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{event_probabilities}.
}
