% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.event_probabilities}
\alias{print.event_probabilities}
\title{Print a short summary for event probabilities}
\usage{
\method{print}{event_probabilities}(x, ...)
}
\arguments{
\item{x}{An object of \code{event_probabilities} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{update_model}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{event_probabilities}.
}
