% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_model_query.R
\name{print.model_query}
\alias{print.model_query}
\title{Print a tightened summary of model queries}
\usage{
\method{print}{model_query}(x, ...)
}
\arguments{
\item{x}{An object of \code{model_query} class.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{model_query}.
}
\seealso{
\code{\link{plot.model_query}} for a graphical display of the same object.
}
\keyword{internal}
