% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{print.nodal_types}
\alias{print.nodal_types}
\title{Print a short summary for causal_model nodal-types}
\usage{
\method{print}{nodal_types}(x, ...)
}
\arguments{
\item{x}{An object of \code{nodal_types} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{make_model} or \code{update_model}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{nodal_types}.
}
