% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.parameter_mapping}
\alias{print.parameter_mapping}
\title{Print a short summary for paramater mapping matrix}
\usage{
\method{print}{parameter_mapping}(x, ...)
}
\arguments{
\item{x}{An object of \code{parameter_mapping} class.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{parameter_mapping}.
}
