% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.parameters}
\alias{print.parameters}
\title{Print a short summary for causal_model parameters}
\usage{
\method{print}{parameters}(x, ...)
}
\arguments{
\item{x}{An object of \code{parameters} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{make_model} or \code{update_model}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{parameters}.
}
