% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{print.parameters_df}
\alias{print.parameters_df}
\title{Print a short summary for a causal_model parameters data-frame}
\usage{
\method{print}{parameters_df}(x, ...)
}
\arguments{
\item{x}{An object of \code{parameters_df} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{make_model} or \code{update_model}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{parameters_df}.
}
