% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.parameters_posterior}
\alias{print.parameters_posterior}
\title{Print a short summary for causal_model parameter posterior distributions}
\usage{
\method{print}{parameters_posterior}(x, ...)
}
\arguments{
\item{x}{An object of \code{parameters_posterior} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{update_model}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{parameters_posterior}.
}
