% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.parameters_prior}
\alias{print.parameters_prior}
\title{Print a short summary for causal_model parameter prior distributions}
\usage{
\method{print}{parameters_prior}(x, ...)
}
\arguments{
\item{x}{An object of \code{parameters_prior} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{set_prior_distribution}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{parameters_prior}.
}
