% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.parents}
\alias{print.parents}
\title{Print a short summary for list of parents}
\usage{
\method{print}{parents}(x, ...)
}
\arguments{
\item{x}{An object of the \code{parents} class, which is a list
containing the names of each parent node in the \code{causal_model} produced using
\code{get_parents}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{parents}.
}
