% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.parents_df}
\alias{print.parents_df}
\title{Print a short summary for a causal_model parents data-frame}
\usage{
\method{print}{parents_df}(x, ...)
}
\arguments{
\item{x}{An object of \code{parents_df} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{make_model} or \code{update_model}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{parents_df}.
}
