% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{print.posterior_event_probabilites}
\alias{print.posterior_event_probabilites}
\title{Print a short summary for event probability distributions}
\usage{
\method{print}{posterior_event_probabilites}(x, ...)
}
\arguments{
\item{x}{An object of \code{posterior_event_probabilities} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{update_model}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{posterior_event_probabilities}.
}
