% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_model.R
\name{print.posterior_event_probabilities}
\alias{print.posterior_event_probabilities}
\title{Print a short summary of posterior_event_probabilities}
\usage{
\method{print}{posterior_event_probabilities}(x)
}
\arguments{
\item{x}{An object of \code{posterior_event_probabilities} class.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{posterior_event_probabilities}.
}
