% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{print.stan_fit}
\alias{print.stan_fit}
\title{Print a short summary for stan fit}
\usage{
\method{print}{stan_fit}(x, ...)
}
\arguments{
\item{x}{An object of \code{stan_fit} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{update_model}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{stan_fit}.
}
