% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.stan_objects}
\alias{print.stan_objects}
\title{Print a short summary for stan_objects}
\usage{
\method{print}{stan_objects}(x, ...)
}
\arguments{
\item{x}{An object of the \code{stan_objects} class, which is a list
containing the data, \code{type_distribution} and  \code{stan_summary}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{stan_objects}.
}
