% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.statement}
\alias{print.statement}
\title{Print a short summary for a causal_model statement}
\usage{
\method{print}{statement}(x, ...)
}
\arguments{
\item{x}{An object of \code{statement} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{make_model} or \code{update_model}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{statement}.
}
