% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{print.type_posterior}
\alias{print.type_posterior}
\title{Print a short summary for causal-type posterior distributions}
\usage{
\method{print}{type_posterior}(x, ...)
}
\arguments{
\item{x}{An object of \code{type_posterior} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{get_type_prob_multiple}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{type_posterior}.
}
