% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{print.type_prior}
\alias{print.type_prior}
\title{Print a short summary for causal-type prior distributions}
\usage{
\method{print}{type_prior}(x, ...)
}
\arguments{
\item{x}{An object of \code{type_prior} class, which is a sub-object of
an object of the \code{causal_model} class produced using
\code{make_model} or \code{update_model}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{type_prior}.
}
