% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_confounds_df.R
\name{prob_par1_given_par2}
\alias{prob_par1_given_par2}
\title{helper to get conditional probability of nodal types}
\usage{
prob_par1_given_par2(par1, par2, nodal_type, P, type_prob)
}
\arguments{
\item{par1}{parameter 1}

\item{par2}{parameter 1}

\item{nodal_type}{vector of nodal types}

\item{P}{P matrix}

\item{type_prob}{vector of type probabilities}
}
\description{
helper to get conditional probability of nodal types
}
