% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{realise_outcome_c}
\alias{realise_outcome_c}
\title{generates realized outcomes for all causal types by sequentially calculating endogenous nodes.}
\usage{
realise_outcome_c(
  d,
  endogenous_nodes,
  parents_list,
  nodal_types_collapsed,
  nodal_types
)
}
\arguments{
\item{d}{a data.frame of causal types passed from within realise_outcomes}

\item{endogenous_nodes}{a character vector with names of endogenous nodes}

\item{parents_list}{a list of parents nodes for each node}

\item{nodal_types_collapsed}{a list of collapsed nodal types}

\item{nodal_types}{a list of uncolpased nodal types}
}
\value{
a data.frame of realized outcomes for each causal type
}
\description{
generates realized outcomes for all causal types by sequentially calculating endogenous nodes.
}
