% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{reduce_lambda}
\alias{reduce_lambda}
\title{Reduce lambda
If lambda is longer than nodel types, because of a model restriction, subset lambda and renormalize}
\usage{
reduce_lambda(model, lambda)
}
\arguments{
\item{model}{a model created by make_model()}

\item{lambda}{a parameter vector possibly longer than the parameter length expected by model}
}
\description{
Reduce lambda
If lambda is longer than nodel types, because of a model restriction, subset lambda and renormalize
}
