% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{reduce_nodal_types}
\alias{reduce_nodal_types}
\title{Restrict a model}
\usage{
reduce_nodal_types(pcm, restrictions)
}
\arguments{
\item{restrictions}{a list of character vectors. Names in list should match dag's variables. Restrictions are specificified by as nodal types.}

\item{dag}{a dag created by make_dag()}
}
\value{
A dag with restrictions and nodal types saved as attributes.
}
\description{
Restrict a model
}
\examples{
XYdag <- make_dag(add_edges(parent = "X", children = c("Y")))
# restrictions can be specified following nodal_types syntax
reduce_nodal_types(dag = XYdag, restrictions(X = "X0", Y = "Y00"))

# or alternatively variable name can be omitted from restriction
reduce_nodal_types(dag = XYdag, restrictions(X = "0", Y = "00"))
}
