% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{reduce_parameters}
\alias{reduce_parameters}
\title{Reduce parameters
If parameters is longer than nodel types, because of a model restriction, subset lambda and renormalize}
\usage{
reduce_parameters(model, parameters = model$parameters)
}
\arguments{
\item{model}{a model created by make_model()}

\item{parameters}{a parameter vector possibly longer than the parameter length expected by model}
}
\description{
Reduce parameters
If parameters is longer than nodel types, because of a model restriction, subset lambda and renormalize
}
