% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{repulsive_force}
\alias{repulsive_force}
\title{internal helper
calculates the repulsive force component used in generating
force directed graph layouts}
\usage{
repulsive_force(distance, k)
}
\description{
internal helper
calculates the repulsive force component used in generating
force directed graph layouts
}
\keyword{internal}
