% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{restrict_causal_types}
\alias{restrict_causal_types}
\title{Reduce causal types}
\usage{
restrict_causal_types(model, restriction)
}
\arguments{
\item{model}{a model created by make_model()}

\item{restriction}{a quoted expressions defining the restriction}
}
\description{
Reduce causal types
}
