% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{restrict_nodal_types}
\alias{restrict_nodal_types}
\title{Restrict a model}
\usage{
restrict_nodal_types(model, restrictions = NULL, dos = NULL)
}
\arguments{
\item{model}{a model created by make_model()}

\item{restrictions}{a list of character vectors specifying nodal types to be removed from the model. Use \code{get_nodal_types} to see syntax.}

\item{do}{a list containing the names of variables and their assigned value.}
}
\value{
A model with restrictions and nodal types saved as attributes.
}
\description{
Restrict causal types. If priors exist prior probabilities are redistributed over remaining types.
}
\examples{
XYmodel <- make_model(add_edges(parent = "X", children = c("Y")))
# restrictions can be specified following nodal_types syntax
restrict_nodal_types(model = XYmodel, restrictions = list(X = "X0", Y = "Y00"))

# or alternatively variable name can be omitted from restriction
restrict_nodal_types(model = XYmodel, restrictions = list(X = "0", Y = "00"))

# A particularly important restriction is the do operator that removes all but one nodal_type from a node
restrict_nodal_types(model = XYmodel, dos = list(X = 1))
restrict_nodal_types(model = XYmodel, dos = list(Y = 0))

# Restrictions can be iteratively applied
my_model <-  restrict_nodal_types(model = XYmodel, restrictions = list(Y = "10"))
my_model <-  restrict_nodal_types(model = my_model, do = list(X = 1))
get_indicator_matrix(my_model)

# Restrictions can be  with wildcards
my_model <-  restrict_nodal_types(model = XYmodel, restrictions = list(Y = "?0"))
get_indicator_matrix(my_model)
}
