% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{restrict_nodal_types_exp}
\alias{restrict_nodal_types_exp}
\title{Reduce nodal types}
\usage{
restrict_nodal_types_exp(model, statement, join_by = "|", keep = FALSE)
}
\arguments{
\item{model}{a model created by make_model()}

\item{keep}{Logical. If `FALSE`, removes and if `TRUE` keeps only causal types specified by \code{restriction}.}

\item{restriction}{a list of character vectors specifying nodal types to be removed from the model. Use \code{get_nodal_types} to see syntax.}
}
\description{
Reduce nodal types
}
