% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_ambiguity_matrices.R
\name{reveal_data}
\alias{reveal_data}
\title{Reveal data}
\usage{
reveal_data(dag)
}
\arguments{
\item{dag}{A dag as created by \code{make_dag}}
}
\value{
revealed_data
}
\description{
Reveal data based on parents' data realizations and types of exogenous variables
}
\details{
\code{reveal_data} starts off by creating types (via \code{\link{gbiq:::get_types}}). It then takes types of endogenous and reveals their outcome based on the value that their parents took. Exogenous variables outcomes correspond to their type.
}
\keyword{internal}
