% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_lambda}
\alias{set_lambda}
\title{Add a true prior distribution}
\usage{
set_lambda(model, lambda = NULL, random = FALSE, average = FALSE)
}
\arguments{
\item{model}{A model generated by \code{make_model}}

\item{lambda}{lambda vector to add to model}

\item{random}{If true, random truth drawn from priors}

\item{average}{If true, prior mean used}
}
\description{
Add a vector of true parameters -- can be use to draw data, rather than true vector
}
\examples{
set_lambda(make_model("X -> Y"), average = TRUE)

}
