% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_parameters}
\alias{set_parameters}
\title{Add a true parameter vector}
\usage{
set_parameters(model, parameters = NULL, type = "prior_mean", ...)
}
\arguments{
\item{model}{A model generated by \code{make_model}}

\item{parameters}{parameter vector to add to model}

\item{type}{An element of ("define", "flat", prior_mean", "posterior_mean", "prior_draw", "posterior_draw"). Indicates type of parameters to set. With type set to \code{define} use arguments to be passed to \code{make_priors}; otherwise \code{flat} sets equal probabilities on each nodal type in each parameter set; \code{prior_mean, prior_draw, posterior_mean, posterior_draw} take parameters as the means or as draws from the prior or posterior.}
}
\description{
Add a vector of "true" parameters to a model.
}
\examples{
model <- set_parameters(make_model("X -> Y"),
          type = "flat")
# Use make_alpha to define specific parameters
model <- set_parameters(make_model("X -> Y"), type = 'define',
                        alphas = list(Y = c(`(Y[X=1] > Y[X=0])` = 3)))
model$parameters
}
