% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmap.R
\name{set_parmap}
\alias{set_parmap}
\title{Set parmap: a matrix mapping from parameters to data types}
\usage{
set_parmap(model)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}
}
\value{
A matrix
}
\description{
Generates and adds parmap to a model
}
\examples{
set_parmap(model = make_model('X->Y'))

}
