% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{set_sampling_args}
\alias{set_sampling_args}
\title{set_sampling_args
From rstanarm (November 1st, 2019)}
\usage{
set_sampling_args(object, user_dots = list(), user_adapt_delta = NULL, ...)
}
\arguments{
\item{object}{The fit}

\item{user_dots}{User commands}

\item{user_adapt_delta}{Adapt delta passed by the user}

\item{...}{further arguments to be passed to stan}
}
\description{
set_sampling_args
From rstanarm (November 1st, 2019)
}
\details{
Set the sampling arguments
}
