% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{simple_nodes}
\alias{simple_nodes}
\title{function to make a parameters_df from nodal types}
\usage{
simple_nodes(n = 5)
}
\arguments{
\item{n}{number of parents node to construct nodal types for}
}
\value{
a vector of nodal types
}
\description{
function to make a parameters_df from nodal types
}
