% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{strategy_statements}
\alias{strategy_statements}
\title{Generate strategy statements given data}
\usage{
strategy_statements(data, strategies)
}
\arguments{
\item{data}{A data frame with one row}

\item{strategies}{A list of strategies where each strategy
is a set of nodes to be observed}
}
\value{
A string
}
\description{
Helper to generate statements of the form "X = 1 & Y = 0"
from realized data on one observation
}
\examples{
 data.frame(X = 1, M = 0, Y = NA) \%>\%
 strategy_statements(list(c("X", "M", "Y"), "X", "Y"))

}
\keyword{helper}
