% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realise_outcomes.R
\name{traverse_dag}
\alias{traverse_dag}
\title{Helper to generate a sub-DAG by recursively traversing a DAG from a given node until a root node or a node with a do-operation is reached}
\usage{
traverse_dag(node, dag, dos = character(0))
}
\arguments{
\item{node}{a string specifying a node in the DAG}

\item{dag}{a named list of character vectors specifying all nodes in the DAG and their respective parents}
}
\value{
a character vector of nodes
}
\description{
Helper to generate a sub-DAG by recursively traversing a DAG from a given node until a root node or a node with a do-operation is reached
}
\keyword{Used}
\keyword{in}
\keyword{internal}
\keyword{realise_outcomes}
