% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{trim_strategies}
\alias{trim_strategies}
\title{Trim strategies}
\usage{
trim_strategies(model, data)
}
\arguments{
\item{model}{A probabilistic causal model created by make_model()}

\item{data}{A data.frame of variables that can take three values: 0, 1, and NA.}
}
\value{
Returns data events without strategies containing no observed data
}
\description{
Trim strategies
}
