% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{type_matrix}
\alias{type_matrix}
\title{Generate type matrix}
\usage{
type_matrix(parent_n)
}
\arguments{
\item{parent_n}{An integer. Number of parents of a given child.}
}
\description{
Generate type matrix
}
