% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_helpers.R
\name{types_to_nodes}
\alias{types_to_nodes}
\title{Wrapper of types_to_nodes_single
Sapply multiple query conditions}
\usage{
types_to_nodes(model, query)
}
\arguments{
\item{model}{A model created by make_model()}

\item{query}{a query}
}
\description{
Wrapper of types_to_nodes_single
Sapply multiple query conditions
}
