% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_to_nodes.R
\name{types_to_nodes_single}
\alias{types_to_nodes_single}
\title{Identify nodes that satisfy a causal type query}
\usage{
types_to_nodes_single(model, query)
}
\arguments{
\item{model}{A model created by make_model()}

\item{query}{a query}
}
\description{
Identify nodes that satisfy a causal type query
}
\keyword{internal}
