% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_helpers.R
\name{types_to_rows}
\alias{types_to_rows}
\title{types_to_rows}
\usage{
types_to_rows(model, query)
}
\arguments{
\item{model}{A model created by \code{make_model()}}

\item{query}{a query}
}
\description{
types_to_rows
}
\keyword{internal}
