% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_to_rows.R
\name{types_to_rows_single}
\alias{types_to_rows_single}
\title{types_to_rows_single}
\usage{
types_to_rows_single(model, query)
}
\arguments{
\item{model}{A model created by make_model()}

\item{query}{a query}
}
\description{
types_to_rows_single
}
