% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodal_types.R
\name{uncollapse_nodal_types}
\alias{uncollapse_nodal_types}
\title{uncollapse nodal types}
\usage{
uncollapse_nodal_types(nodal_types)
}
\arguments{
\item{nodal_types}{A list of nodal types.}
}
\value{
A \code{list} containing nodes with nodal types in data.frame form
}
\description{
uncollapse nodal types
}
\examples{

model <- make_model('X -> K -> Y')
(nodal_types <- get_nodal_types(model , collapse = FALSE))
CausalQueries:::uncollapse_nodal_types(nodal_types)
}
\keyword{internal}
