% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_restrictions.R
\name{unpack_wildcard}
\alias{unpack_wildcard}
\title{Unpack a wild card}
\usage{
unpack_wildcard(x)
}
\arguments{
\item{x}{A character. A nodal type containing one or more wildcard characters '.' to be unpacked.}
}
\description{
Unpack a wild card
}
