% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{var_in_query}
\alias{var_in_query}
\title{List of nodes contained in query}
\usage{
var_in_query(model, query)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{query}{An expression in string format.}
}
\description{
List of nodes contained in query
}
