% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_type.R
\name{add_dots}
\alias{add_dots}
\title{Helper to fill in missing do operators in causal expression}
\usage{
add_dots(q, model)
}
\arguments{
\item{q}{a causal query}

\item{model}{a model}
}
\description{
Helper to fill in missing do operators in causal expression
}
\examples{
model <- make_model("X -> Y <- M")
gbiqq:::add_dots(Y[X=1], model)
}
