% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type_helpers.R
\name{add_wildcard}
\alias{add_wildcard}
\title{Adds a wildcard for every missing parent}
\usage{
add_wildcard(node, statement, parents, missing_parents)
}
\arguments{
\item{node}{A character string. The quoted name of a node.}

\item{statement}{A character string. A quoted causal statement.}

\item{parents}{A vector of characters. The \code{node}'s parents}

\item{missing_parents}{A vector of characters.  The \code{node}'s missing parents}
}
\value{
A causal query expression with all parents nodes set to either 0, 1 or wildcard '.'
}
\description{
Adds a wildcard for every missing parent
}
\keyword{internal}
