% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{all_data_types}
\alias{all_data_types}
\title{All data types}
\usage{
all_data_types(model)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}
}
\description{
Creates dataframe with all data types, including NA types possible from a model
}
\examples{
all_data_types(make_model("X -> Y"))
}
