% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_helpers.R
\name{any_discrepancies}
\alias{any_discrepancies}
\title{Check for discrepancies
Used in make_alphas}
\usage{
any_discrepancies(alphas, translated_alphas)
}
\arguments{
\item{alphas}{alphas provided by the user in set_priors/make_priors}

\item{translated_alphas}{alphas that were expressed as a causal query.}
}
\description{
Check for discrepancies
Used in make_alphas
}
