% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_and_estimands.R
\name{calculate_estimand}
\alias{calculate_estimand}
\title{Calculate estimand}
\usage{
calculate_estimand(model, do1, query1, do2 = NULL, query2 = NULL,
  subset1 = TRUE, subset2 = TRUE, aggregation = function(a, b) a - b,
  sims = 100, lambda = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{query}{A logic operation on variables, in quotes. For example "Y==1"}

\item{subset}{An optional logic condition on variabless, in quotes.. For example "X==0"}
}
\description{
Ask a question of a model
}
\examples{
model <- make_model(add_edges(parent = "X", children = c("Y")))
model <- reduce_nodal_types(model = model, restrictions = list(Y = "Y10"))
estimand <- calculate_estimand (model = model,
										do1 = list(X = 0),
										query1 = "Y==1",
										do2 = list(X = 1),
										query2 = "Y==1",
										aggregation = function(q1,q2) q2-q1,
										sims = 300)
summary(estimand)
}
