% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_and_estimands.R
\name{calculate_multiple_estimands}
\alias{calculate_multiple_estimands}
\title{Calculate multipl estimands}
\usage{
calculate_multiple_estimands(model, queries = list(NULL),
  subsets = list(TRUE), posteriors = list(FALSE),
  estimand_labels = NULL, stats = c(mean = mean, sd = sd))
}
\arguments{
\item{model}{A  model}

\item{posterior}{if true use a posterior distribution, otherwise use the prior}

\item{subset}{quoted expression evaluates to logical statement. subset allows estimand to be conditioned on *observational* distribution.}
}
\description{
Calculated from a prior or posterior distribution
}
\examples{
model <- make_model("X" \%->\% "Y") \%>\%
           set_prior_distribution()

calculate_multiple_estimands(
      model,
      queries = list(ATE = "Y[X=1] - Y[X=0]", Share_positive = "Y[X=1] > Y[X=0]"))

}
