% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalQueries-package.R
\docType{package}
\name{CausalQueries-package}
\alias{CausalQueries-package}
\title{'CausalQueries'}
\description{
'CausalQueries' is a package that lets you declare binary causal models,
update beliefs about causal types given data and calculate arbitrary estimands.
Model definition makes use of dagitty syntax. Updating is implemented in 'stan'.
}
