% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalQueries-package.R
\docType{package}
\name{CausalQueries-package}
\alias{CausalQueries-package}
\title{CausalQueries: A package for creating and updating  binary probabilistic causal models}
\description{
CausalQueries: A package for creating and updating  binary probabilistic causal models
}
