% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalQueries-package.R
\docType{package}
\name{CausalQueries-package}
\alias{CausalQueries-package}
\title{'CausalQueries'}
\description{
'CausalQueries' is a package that lets you generate binary causal models,
update over models given data and calculate arbitrary causal queries.
Model definition makes use of dagitty syntax.
Updating is implemented in 'stan'.
}
