% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalQueries-package.R
\docType{_PACKAGE}
\name{CausalQueries}
\alias{CausalQueries}
\title{'CausalQueries'}
\description{
'CausalQueries' is a package that lets users generate binary causal models,
update over models given data, and calculate arbitrary causal queries.
Model definition makes use of dagitty type syntax.
Updating is implemented in 'stan'.
}
