% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{check_params}
\alias{check_params}
\title{Check parameters sum to 1 in paramset and normalize}
\usage{
check_params(parameters, param_set_names, warning = FALSE,
  model = NULL)
}
\description{
Check parameters sum to 1 in paramset and normalize
}
