% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{check_params}
\alias{check_params}
\title{Check parameters sum to 1 in paramset; normalize if needed; add names if needed}
\usage{
check_params(parameters, param_set_names, warning = FALSE,
  model = NULL)
}
\arguments{
\item{parameters}{A vector of probabilities, summing to 1 within each parameter set.}

\item{param_set_names}{Names of parameter sets. Param names begin with param_set_names followed by a period. For instaicen "X.X0".}

\item{warning}{Logical whether to print warning (if any) in console.}

\item{model}{A model object generated by \code{make_model()}.}
}
\description{
Check parameters sum to 1 in paramset; normalize if needed; add names if needed
}
\examples{
\dontrun{
check_params(rep(1,6), param_set_names = c("X", "Y"), warning = TRUE)
check_params(rep(1,6), param_set_names = c("X", "Y"), model = make_model("X->Y"))
check_params(rep(1,6), param_set_names = c("X", "Y"), model = make_model("X->Y"), warning = TRUE)
}
}
