% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{check_priors}
\alias{check_priors}
\title{Check parameters sum to 1 in paramset and normalize}
\usage{
check_priors(priors, model = NULL)
}
\description{
Check parameters sum to 1 in paramset and normalize
}
