% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_params.R
\name{clean_params}
\alias{clean_params}
\title{Check parameters sum to 1 in paramset; normalize if needed; add names if needed}
\usage{
clean_params(parameters_df, warning = TRUE)
}
\arguments{
\item{parameters_df}{dataframe, first generated by make_model}

\item{warning}{Logical whether to print warning (if any) in console.}
}
\description{
Check parameters sum to 1 in paramset; normalize if needed; add names if needed
}
\examples{
model <- make_model("X->Y")
model$parameters_df$parameters <- 1:6
clean_params(model$parameters_df, warning = TRUE)
}
