% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_params.R
\name{clean_params}
\alias{clean_params}
\title{Check parameters sum to 1 in param_set; normalize if needed; add
names if needed}
\usage{
clean_params(parameters_df, warning = TRUE)
}
\arguments{
\item{parameters_df}{A \code{data.frame}. This object is first generated by
\code{\link{make_model}}.}

\item{warning}{Logical. Whether to print warning (if any) in console.
Defaults to TRUE}
}
\value{
A parameters \code{data.frame}  with names and where parameters
  sum to 1.
}
\description{
Check parameters sum to 1 in param_set; normalize if needed; add
names if needed
}
\examples{
\donttest{
model <- make_model('X->Y')
model$parameters_df$param_value <- 1:6
CausalQueries:::clean_params(model$parameters_df, warning = TRUE)
}
}
\keyword{internal}
