% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{collapse_data}
\alias{collapse_data}
\title{Make data compact with data as first argument}
\usage{
collapse_data(data, model, remove_family = TRUE)
}
\arguments{
\item{data}{A data.frame.}

\item{model}{A model}
}
\description{
Make data compact with data as first argument
}
