% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodal_types.R
\name{collapse_nodal_types}
\alias{collapse_nodal_types}
\title{collapse nodal types}
\usage{
collapse_nodal_types(nodal_types, include_node_names = FALSE)
}
\arguments{
\item{nodal_types}{A list of nodal types.}

\item{include_node_names}{Logical, if TRUE returns names X0, X1; otherwise returns 0, 1}
}
\description{
collapse nodal types
}
