% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodal_types.R
\name{collapse_nodal_types}
\alias{collapse_nodal_types}
\title{collapse nodal types}
\usage{
collapse_nodal_types(nodal_types, include_node_names = FALSE)
}
\arguments{
\item{nodal_types}{A list of nodal types.}

\item{include_node_names}{Logical, if TRUE returns names X0, X1; otherwise returns 0, 1}
}
\value{
A \code{list} containing nodes with nodal types in a vector form.
}
\description{
collapse nodal types
}
\examples{

model <- make_model('X -> K -> Y')
(nodal_types <- get_nodal_types(model , collapse = FALSE))
CausalQueries:::collapse_nodal_types(nodal_types )
}
\keyword{internal}
