% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_helpers.R
\name{complements}
\alias{complements}
\title{Make statement for complements}
\usage{
complements(X1, X2, Y)
}
\arguments{
\item{X1}{input node 1}

\item{X2}{input node 2}

\item{Y}{outcome node}
}
\value{
A character statement of class statement
}
\description{
Generate a statement for X1, X1 complement each other in the production of Y
}
\examples{
complements("A", "B", "W")

}
\seealso{
Other statements: \code{\link{decreasing}},
  \code{\link{increasing}}, \code{\link{interacts}},
  \code{\link{substitutes}}, \code{\link{te}}
}
\concept{statements}
