% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_and_estimands.R
\name{conditional_inferences}
\alias{conditional_inferences}
\title{Conditional inferences}
\usage{
conditional_inferences(model, lambda, query, given = NULL)
}
\description{
Calculate estimands condition on observed data (currently for single case process tracing) together with data realization probabilities
}
\examples{
model <- make_model("X" \%->\% "Y")
model <- set_lambda(model, average = TRUE)
conditional_inferences(model, model$lambda, query = "Y[X=1]>Y[X=0]")
}
