% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_confounds.R
\name{continue_names}
\alias{continue_names}
\title{Continue names}
\usage{
continue_names(x, split = "_")
}
\arguments{
\item{x}{A vector with strings of the form c('x_1', 'x_2')}

\item{split}{String on which to split. String should appear only once in each vector entry.}
}
\description{
Slightly hacky function to continue param_set names in a sequence
}
\examples{
x <- c('S_3', 'S_3', 'S_5')
gbiqq:::continue_names(x)
}
