% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{data_strategy}
\alias{data_strategy}
\title{Data Strategy}
\usage{
data_strategy(model, parameters = NULL, n, vars = list(NULL),
  probs = list(NULL), ms = NULL, subsets = list(NULL))
}
\description{
Data Strategy
}
\examples{
 # A strategy in which X, Y are observed for sure and M is observed with 50\% probabilit for X=1, Y=0 cases
model <- make_model("X -> M -> Y")
model <- set_parameters(model, type = "flat")
data_strategy(
   model,
   n = 8,
   vars = list(c("X", "Y"), "M"),
   probs = list(1, .5),
   subsets = list(NULL, "X==1 & Y==0"))
}
