% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{data_strategy}
\alias{data_strategy}
\title{Data Strategy}
\usage{
data_strategy(model, lambda = NULL, n, vars = list(NULL),
  probs = list(NULL), ms = NULL, subsets = list(NULL))
}
\description{
Data Strategy
}
\examples{
model <- make_model("X" \%->\% "M", "M" \%->\% "Y")
data_strategy(
   model,
   n = 8,
   vars = list(c("X", "Y"), "M"),
   probs = list(1, .5),
   subsets = list(NULL, "X==1 & Y==0"))
}
