% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{data_type_names}
\alias{data_type_names}
\title{Data type names}
\usage{
data_type_names(model, data)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by
\code{\link{make_model}}.}

\item{data}{A \code{data.frame}.  Data of nodes that can take three values:
0, 1, and NA. In long form as generated by \code{\link{make_events}}}
}
\value{
A vector of strings of data types
}
\description{
Provides names to data types
}
\examples{
model <- make_model('X -> Y')
data <- make_data(model, n = 2)
data_type_names(model, data)
}
\keyword{internal}
