% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{default_stan_control}
\alias{default_stan_control}
\title{default_stan_control}
\usage{
default_stan_control(adapt_delta = NULL, max_treedepth = 15L)
}
\arguments{
\item{adapt_delta}{A double between 0 and 1. It determines \code{adapt_delta}}

\item{max_treedepth}{A positive integer. It determines  \code{maximum_tree_depth}}
}
\description{
default_stan_control
}
\details{
Sets controls to default unless otherwise specified.
}
\keyword{internal}
