% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{define_priors}
\alias{define_priors}
\title{Define priors}
\usage{
define_priors(dag = NULL, prior_distribution = "uniform",
  alphas = NULL)
}
\arguments{
\item{dag}{A dag created by make_dag()}

\item{nodal_types}{}
}
\description{
Define priors
}
