% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_data}
\alias{draw_data}
\title{Draw compact data}
\usage{
draw_data(model, n = 1, w = NULL, P = NULL, A = NULL,
  lambda = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{n}{Number of observations}

\item{w}{Vector of event probabilities}

\item{P}{Optional parameter matrix: not required but may be provided to avoide repeated computation for simulations}

\item{A}{Optional ambiguity matrix: not required but may be provided to avoide repeated computation for simulations}

\item{lambda}{A specific parameter vector, lambda, may be provided, otherwise lambda is drawn from priors}
}
\description{
Draw compact data
}
\examples{
model <- make_model(add_edges(parent = "X", children = c("Y")))
draw_data(model = model)
}
