% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_data_events}
\alias{draw_data_events}
\title{Draw compact data}
\usage{
draw_data_events(model, n = 1, w = NULL, P = NULL, A = NULL,
  parameters = NULL, using = NULL)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{n}{An integer. Number of observations.}

\item{w}{A numeric vector of event probabilities.}

\item{P}{Parameter matrix. (Optional). May be provided to avoid repeated computation for simulations.}

\item{A}{Ambiguity matrix. (Optional). May be provided to avoid repeated computation for simulations.}

\item{parameters}{A numeric vector. Values of parameters may be specified. By default, it is drawn from priors.}

\item{using}{String indicating whether to use "priors", "posteriors" or "parameters".}
}
\value{
A data frame of events
}
\description{
Draw compact data
}
\examples{
model <- make_model("X -> Y")
draw_data_events(model = model)
}
