% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_data_events}
\alias{draw_data_events}
\title{Draw compact data}
\usage{
draw_data_events(model, n = 1, w = NULL, P = NULL, A = NULL,
  parameters = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{n}{Number of observations}

\item{w}{Vector of event probabilities}

\item{P}{Optional parameter matrix: not required but may be provided to avoide repeated computation for simulations}

\item{A}{Optional ambiguity matrix: not required but may be provided to avoide repeated computation for simulations}

\item{parameters}{A specific parameter vector, parameters, may be provided, otherwise parameters is drawn from priors}
}
\description{
Draw compact data
}
\examples{
model <- make_model("X -> Y")
draw_data_events(model = model)
}
