% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_event_prob}
\alias{draw_event_prob}
\title{Draw event probabilities}
\usage{
draw_event_prob(model, P = NULL, A = NULL, parameters = NULL,
  type_prob = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{P}{Parameter matrix, not required but may be provided to avoid repeated computation for simulations}

\item{A}{Ambiguity matrix, not required but may be provided to avoid repeated computation for simulations}

\item{parameters}{A specific parameter vector, parameters; if not  provided,  parameters is drawn from priors}
}
\description{
Draw event probabilities
}
\examples{
model <- make_model("X -> Y")
draw_event_prob(model = model)
}
