% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_lambda}
\alias{draw_lambda}
\title{Draw parameters}
\usage{
draw_lambda(model, using = "priors")
}
\arguments{
\item{model}{A model created by make_model()}

\item{using}{String indicating whether to use `priors`, `posteriors` or `parameters`}
}
\description{
Draw parameters
}
\examples{
draw_lambda(model = model)
}
