% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_lambda}
\alias{draw_lambda}
\title{Draw parameters
Default behavior takes a random draw of parameters from priors; can also draw from posteriors, or return pre defined parameters.}
\usage{
draw_lambda(model, using = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{using}{String indicating whether to use `priors`, `posteriors` or `parameters`}
}
\description{
Draw parameters
Default behavior takes a random draw of parameters from priors; can also draw from posteriors, or return pre defined parameters.
}
\examples{
draw_lambda(model = model)
}
