% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_lambda}
\alias{draw_lambda}
\title{Draw parameters}
\usage{
draw_lambda(model)
}
\arguments{
\item{model}{A model created by make_model()}
}
\description{
Draw parameters
}
\examples{
draw_lambda(model = model)
}
