% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{draw_type_prob}
\alias{draw_type_prob}
\title{Draw type probabilities}
\usage{
draw_type_prob(model, P = NULL, lambda = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{P}{Parameter matrix, not required but may be provided to avoide repeated computation for simulations}

\item{lambda}{A specific parameter vector, lambda, may be provided, otherwise lambda is drawn from priors}
}
\description{
Draw type probabilities
}
\examples{
model <- make_model(add_edges(parent = "X", children = c("Y")))
draw_type_prob(model = model)
}
