% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_type_prob.R
\name{draw_type_prob}
\alias{draw_type_prob}
\title{Draw type probabilities}
\usage{
draw_type_prob(model, P = NULL, parameters = NULL, using = NULL)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{P}{A matrix. Parameter matrix. Not required but may be provided to avoid repeated computation for simulations.}

\item{parameters}{A numeric vector. Optionally specify parameters. By default, parameters is drawn from `using` argument (either from priors, posteriors, or from model$parameters)}

\item{using}{A string taking value "priors", "posteriors" or "parameters".}
}
\description{
Draws probability of vector of causal types given a single realization of parameters, possibly drawn from model priors.
}
\examples{
draw_type_prob(model = make_model("X->Y"))

}
