% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{expand_data}
\alias{expand_data}
\title{Expand compact data object to data frame}
\usage{
expand_data(data_events, model)
}
\arguments{
\item{data_events}{A compact data frame compatible with \code{model}.}

\item{model}{A model object generated by \code{make_model()}.}
}
\description{
Expand compact data object to data frame
}
\examples{
library(dplyr)
model <- make_model("X->M->Y")
draw_data_events(model, n = 5) \%>\%
  expand_data(model)
draw_data_events(model, n = 0) \%>\%
  expand_data(model)

}
