% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{expand_data}
\alias{expand_data}
\title{Expand compact data object to data frame}
\usage{
expand_data(data_events, model)
}
\description{
#' @inheritParams gbiqq_internal_inherit_params
}
\examples{
model <- make_model('X->M->Y')
simulate_events(model, n = 5) \%>\%
  expand_data(model)
simulate_events(model, n = 0) \%>\%
  expand_data(model)

}
