% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_query_to_nodal_type.R
\name{expand_nodal_expression}
\alias{expand_nodal_expression}
\title{Helper to expand nodal expression}
\usage{
expand_nodal_expression(model, query, node, join_by = "|")
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}

\item{query}{A character string. An expression defining nodal types to interrogate \code{\link{reveal_outcomes}}}

\item{node}{A character string. The quoted name of a node.}

\item{join_by}{A logical operator. Used to connect causal statements: \emph{AND} ('&') or \emph{OR} ('|'). Defaults to '|'.}
}
\description{
Helper to expand nodal expression
}
