% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expand_wildcard}
\alias{expand_wildcard}
\title{Expand wildcard}
\usage{
expand_wildcard(to_expand, join_by = "|")
}
\arguments{
\item{to_expand}{A character vector of length 1L.}

\item{join_by}{A character vector of length 1L.}
}
\description{
Expand statement containing wildcard
}
\examples{
expand_wildcard("(Y[X=1, M=.] > Y[X=1, M=.])")

}
