% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbiqq.R
\name{gbiqq}
\alias{gbiqq}
\title{Fit stan model}
\usage{
gbiqq(dag, data, lambda_priors = NULL, P = NULL, ...)
}
\arguments{
\item{dag}{A dag as created by \code{make_dag}}

\item{data}{A data frame with observations}

\item{lambdas_prior}{A vector containg priors for lambda}
}
\description{
Fit stan model
}
