% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbiqq.R
\name{gbiqq}
\alias{gbiqq}
\title{Fit stan model}
\usage{
gbiqq(model, data, ...)
}
\arguments{
\item{data}{A data frame with observations}

\item{dag}{A dag as created by \code{make_dag}}
}
\description{
Takes a model and data and returns a model object with data attached and a posterior model
}
