% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbiqq.R
\name{gbiqq}
\alias{gbiqq}
\title{Fit stan model}
\usage{
gbiqq(model, data, ...)
}
\arguments{
\item{model}{A model as created by \code{make_model}}

\item{data}{A data frame with observations}
}
\description{
Takes a model and data and returns a model object with data attached and a posterior model
}
