% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbiqq.R
\name{gbiqq}
\alias{gbiqq}
\title{Fit stan model}
\usage{
gbiqq(model, data, stan_model = NULL, ...)
}
\arguments{
\item{model}{A causal model as created by \code{make_model}.}

\item{data}{A data frame with observations}

\item{stan_model}{A fitted stan model. If not provided a gbiqq model is compiled from stan file "simplexes.stan"}

\item{...}{Options passed onto \code{rstan::stan} call.}
}
\description{
Takes a model and data and returns a model object with data attached and a posterior model
}
