% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbiqq_designer.R
\name{gbiqq_designer}
\alias{gbiqq_designer}
\title{vars = list(c("X", "Y"), "M"),
   probs = list(1, .5),
   subsets = list(NULL, "X==1 & Y==0"))}
\usage{
gbiqq_designer(dag = make_model("X" \%->\% "Y"), restrictions = NULL,
  priors = "uniform", data_strategy = list(n = 1, vars = list(NULL),
  probs = list(NULL), ms = NULL, subsets = list(NULL)),
  estimand = "Y[X=1] - Y[X=0]")
}
\description{
vars = list(c("X", "Y"), "M"),
   probs = list(1, .5),
   subsets = list(NULL, "X==1 & Y==0"))
}
