% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbiqq_designer.R
\name{gbiqq_designer}
\alias{gbiqq_designer}
\title{Integrate gbiqq with DeclareDesign}
\usage{
gbiqq_designer(model = make_model("X" \%->\% "Y"), restrictions = NULL,
  priors = "uniform", lambda = NULL, data_strat = list(n = 1, vars =
  list(NULL), probs = list(NULL), ms = NULL, subsets = list(NULL)),
  queries = list(ATE = "Y[X=1] - Y[X=0]"))
}
\arguments{
\item{model}{A model created by make_model()}
}
\description{
Define a model, estimand, and answer strategy
}
\examples{
my_design <- gbiqq_designer(
model <- make_model("X -> Y")
  lambda = c(.5, .5, .1, .7, .1, .1),  # Truth
  data_strat = list(n = 4, vars = list(NULL), probs = list(NULL), ms = NULL, subsets = list(NULL)),
  queries = list(ATE = "Y[X=1] - Y[X=0]"))
  draw_data(my_design)
  draw_estimands(my_design)
  draw_estimates(my_design)
}
