% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ambiguities_matrix.R
\name{get_ambiguities_matrix}
\alias{get_ambiguities_matrix}
\title{Get ambiguities matrix}
\usage{
get_ambiguities_matrix(model)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}
}
\value{
A data frame. Types (rows) corresponding to possible data realizations (columns).

A matrix representing ambiguities.
}
\description{
Get matrix that maps types to data realizations.
}
\examples{
model <- make_model("X -> Y")
get_ambiguities_matrix(model = model)

}
