% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ambiguities_matrix.R
\name{get_ambiguities_matrix}
\alias{get_ambiguities_matrix}
\title{Get ambiguities matrix}
\usage{
get_ambiguities_matrix(model)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by
\code{\link{make_model}}.}
}
\value{
A \code{data.frame}. Causal types (rows) corresponding to
possible data realizations (columns).
}
\description{
Return ambiguities matrix if it exists; otherwise calculate
it assuming no confounding.The ambiguities matrix maps from causal types
into data types.
}
