% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_ambiguity_matrices.R
\name{get_ambiguity_matrix}
\alias{get_ambiguity_matrix}
\title{Get ambiguity matrix}
\usage{
get_ambiguity_matrix(dag)
}
\arguments{
\item{dag}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Get matrix that maps types to data realizations
}
\examples{
# ADD_EXAMPLES_HERE
}
