% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_causal_types.R
\name{get_causal_types}
\alias{get_causal_types}
\title{Get causal types}
\usage{
get_causal_types(model)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}
}
\value{
A data frame.
}
\description{
Create a data frame with types produced from all combinations of possible data produced by a DAG.
}
