% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_deprecate.R
\name{get_children_of_exogenous}
\alias{get_children_of_exogenous}
\title{Get children of exogenous variables}
\usage{
get_children_of_exogenous(model)
}
\arguments{
\item{A}{probabilistic causal model created by make_model()}
}
\value{
A list of the children of exogenous variables
}
\description{
Get children of exogenous variables
}
