% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan.R
\name{get_data_events}
\alias{get_data_events}
\title{Summarize data events}
\usage{
get_data_events(data, model)
}
\arguments{
\item{data}{A data.frame of variables that can take three values: 0, 1, and NA}

\item{model}{A dag created by make_dag()}
}
\value{
A vector of data events
}
\description{
Take a dataframe and return compact dataframe of event types and strategies.
}
