% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_and_estimands.R
\name{get_data_probs}
\alias{get_data_probs}
\title{Get data probabilities}
\usage{
get_data_probs(model, data)
}
\arguments{
\item{model}{A  model}

\item{data}{Data in long format}
}
\description{
Takes in a matrix of possible (single case) observations and returns the probability of each
}
\examples{
model <- make_model("X" \%->\% "Y")
data <- simulate_data(model, n = 4)
get_data_probs(model, data)
}
