% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_deprecate.R
\name{get_endogenous_vars}
\alias{get_endogenous_vars}
\title{Get endogenous variables in a DAG}
\usage{
get_endogenous_vars(model)
}
\arguments{
\item{A}{probabilistic causal model created by make_model()}
}
\value{
A vector of endogenous variables
}
\description{
Get endogenous variables in a DAG
}
