% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_and_estimands.R
\name{get_estimands}
\alias{get_estimands}
\title{Calculate multiple estimands}
\usage{
get_estimands(model, lambda = NULL, queries = list(NULL),
  subsets = list(TRUE), posteriors = list(FALSE),
  estimand_labels = NULL, stats = NULL)
}
\arguments{
\item{model}{A  model}

\item{lambda}{A  true parameter vector, if available}

\item{posterior}{if true use a posterior distribution, otherwise use the prior}

\item{subset}{quoted expression evaluates to logical statement. subset allows estimand to be conditioned on *observational* distribution.}
}
\description{
Calculated from a parameter vector, from a prior or from a posterior distribution
}
\examples{
model <- make_model("X -> Y") \%>\%
           set_prior_distribution(n_draws = 10000)

get_estimands(
      model,
      queries = list(ATE = "Y[X=1] - Y[X=0]",
                     Share_positive = "Y[X=1] > Y[X=0]"))

}
