% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_and_estimands.R
\name{get_estimands}
\alias{get_estimands}
\title{Generate estimands dataframe}
\usage{
get_estimands(model, parameters = NULL, queries = list(NULL),
  subsets = list(TRUE), using = list(FALSE), stats = NULL,
  digits = 3, n_draws = 4000)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{parameters}{A numeric vector of  true parameters, if available.}

\item{queries}{A string. Query on potential otcomes such as "Y[X=1] - Y[X=0]".}

\item{subsets}{A quoted expression that evaluates to a logical statement. Allows estimand to be conditioned on *observational* distribution.}

\item{using}{String indicating whether to use `priors`, `posteriors` or `parameters`.}

\item{stats}{Functions be applied to estimand distribution. If `NULL`, defaults to mean and standard deviation.}

\item{digits}{An integer. Decimal digits in output table.}

\item{n_draws}{An integer. Number of draws.}
}
\description{
Calculated from a parameter vector, from a prior or from a posterior distribution
}
\examples{
model <- make_model("X -> Y") \%>\%
           set_prior_distribution(n_draws = 10000)

get_estimands(
      model,
      queries = list(ATE = "Y[X=1] - Y[X=0]",
                     Share_positive = "Y[X=1] > Y[X=0]"),
      using = c("parameters", "priors"))

get_estimands(
      model,
      queries = list(ATE = "Y[X=1] - Y[X=0]",
                     Share_positive = "Y[X=1] > Y[X=0]"),
      using = "priors")

get_estimands(
      model,
      queries = list(ATE = "Y[X=1] - Y[X=0]"),
      using = list("priors", "parameters"),
      digits = 3)

get_estimands(
      model,
      using = "priors",
      queries = list(Is_B = "Y[X=1] > Y[X=0]"),
      subsets = list(TRUE, "Y==1 & X==1", "Y==0 & X==1"),
      digits = 3)

get_estimands(
      model,
      using = "parameters",
      queries = list(Is_B = "Y[X=1] > Y[X=0]"),
      subsets = list(TRUE, "Y[X=1]==1", "Y==1"),
      digits = 3)
}
