% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_dags.R
\name{get_exogenous_vars}
\alias{get_exogenous_vars}
\title{Get exogenous variables in a DAG}
\usage{
get_exogenous_vars(dag)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
A vector of exogenous variables
}
\description{
Get exogenous variables in a DAG
}
