% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_ambiguity_matrices.R
\name{get_expanded_types}
\alias{get_expanded_types}
\title{Get expanded types}
\usage{
get_expanded_types(pcm)
}
\arguments{
\item{pcm}{A dag as created by \code{make_dag}}
}
\value{
types
}
\description{
Create a data frame with types produced from all combinations of possible data produce by a dag.
}
