% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguities.R
\name{get_indicator_matrix}
\alias{get_indicator_matrix}
\title{Get indicator matrix}
\usage{
get_indicator_matrix(model)
}
\arguments{
\item{model}{A dag created by make_dag()}
}
\description{
Get indicator matrix
}
