% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan_using_dags.R
\name{get_indicator_matrix}
\alias{get_indicator_matrix}
\title{Get indicator matrix}
\usage{
get_indicator_matrix(dag)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\description{
Get indicator matrix
}
