% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{get_likelihood_helpers}
\alias{get_likelihood_helpers}
\title{get_likelihood_helpers}
\usage{
get_likelihood_helpers(model)
}
\arguments{
\item{model}{A probabilistic causal model created by make_model()}
}
\value{
Returns indices and ambiguity matrix
}
\description{
Produce the indices and ambiguity matrix needed for the
}
