% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{get_max_possible_data}
\alias{get_max_possible_data}
\title{Get data frame with all possible data combinations
# FLAG -- this is a redundant function called to make ambiguities matrix}
\usage{
get_max_possible_data(model)
}
\arguments{
\item{model}{A model created by make_model()}
}
\value{
A data frame
}
\description{
Get data frame with all possible data combinations
# FLAG -- this is a redundant function called to make ambiguities matrix
}
