% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_possible_data.R
\name{get_max_possible_data}
\alias{get_max_possible_data}
\title{Get data frame with all possible data combinations}
\usage{
get_max_possible_data(model)
}
\arguments{
\item{model}{A model created by make_model()}
}
\value{
A data frame
}
\description{
Get data frame with all possible data combinations
}
