% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_dags.R
\name{get_max_possible_data}
\alias{get_max_possible_data}
\title{Get data frame with all possible data combinations}
\usage{
get_max_possible_data(dag)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
A data frame
}
\description{
Get data frame with all possible data combinations
}
