% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_dags.R
\name{get_n_endogenous_types}
\alias{get_n_endogenous_types}
\title{Get number of endogenous types}
\usage{
get_n_endogenous_types(dag)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
A list of the numbers of endogenous types
}
\description{
Get number of endogenous types
}
