% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_models.R
\name{get_nodal_types}
\alias{get_nodal_types}
\title{Get list of types for variables in a DAG}
\usage{
get_nodal_types(model, collapse = TRUE)
}
\arguments{
\item{A}{probabilistic causal model created by make_model()}
}
\value{
A list of parents in a DAG
}
\description{
Get list of types for variables in a DAG
}
\examples{
\dontrun{
dag <- make_dag(
 add_edges(parent = "X",children = c("K","Y")),
 add_edges(parent = "K",children = "Y")
)

get_types(dag)
}

}
