% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodal_types.R
\name{get_nodal_types}
\alias{get_nodal_types}
\title{Get list of types for variables in a DAG}
\usage{
get_nodal_types(model, collapse = TRUE)
}
\arguments{
\item{model}{A model created by make_model()}
}
\value{
A list of parents in a DAG
}
\description{
As type labels are hard to interpret for large models, the type list includes an attribute to help interpret them. See  \code{attr(types, interpret)}
}
\examples{
\dontrun{
model <- make_model(
 add_edges(parent = "X",children = c("K","Y")),
 add_edges(parent = "K",children = "Y")
)

get_nodal_types(model)
}

}
