% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_ambiguity_matrices.R
\name{get_nodal_types}
\alias{get_nodal_types}
\title{Get nodal types}
\usage{
get_nodal_types(dag)
}
\arguments{
\item{dag}{A dag as created by \code{make_dag}}
}
\value{
types
}
\description{
Get nodal types
}
\keyword{internal}
