% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_models.R
\name{get_nodal_types_dep}
\alias{get_nodal_types_dep}
\title{Get nodal types
Nodal types are created by concatening variables and their possible data. Used for labeling ambiguities matrix.}
\usage{
get_nodal_types_dep(model)
}
\arguments{
\item{model}{A dag as created by \code{make_dag}}
}
\value{
types
}
\description{
Get nodal types
Nodal types are created by concatening variables and their possible data. Used for labeling ambiguities matrix.
}
