% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_type_prob.R
\name{get_param_dist}
\alias{get_param_dist}
\title{Get a distribution of model parameters}
\usage{
get_param_dist(model, using, n_draws = 4000)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{using}{A string taking value "priors", "posteriors" or "parameters".}

\item{n_draws}{Number of draws from prior}
}
\description{
Using parameters, priors, or posteriors
}
\examples{
get_param_dist(model = make_model("X->Y"), using = "priors", n_draws = 4)
get_param_dist(model = make_model("X->Y"), using = "parameters")
}
