% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_parameter_names}
\alias{get_parameter_names}
\title{get_parameter_names}
\usage{
get_parameter_names(P = NULL, model = NULL)
}
\arguments{
\item{P}{P matrix}

\item{model}{A model}
}
\description{
Parameter names taken from P matrix or model if no P matrix provided
}
