% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_parameter_names}
\alias{get_parameter_names}
\title{Get parameter names}
\usage{
get_parameter_names(model, include_paramset = TRUE)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by
\code{\link{make_model}}.}

\item{include_paramset}{Logical. Whether to include the param set
prefix as part of the name.}
}
\value{
A character vector with the names of the parameters in the model
}
\description{
Parameter names taken from \code{P} matrix or model if no
\code{P}  matrix provided
}
\examples{

get_parameter_names(make_model('X->Y'))
}
