% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters}
\usage{
get_parameters(model, ...)
}
\arguments{
\item{model}{A model object generated by make_model().}

\item{...}{Arguments taken by set_parameters}
}
\description{
Get parameters
}
