% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters}
\usage{
get_parameters(model)
}
\arguments{
\item{model}{A model object generated by make_model().}
}
\description{
Extracts parameters as a namesd vector
}
\examples{
get_parameters(make_model("X -> Y"))
}
