% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parents.R
\name{get_parents}
\alias{get_parents}
\title{Get list of parents of all nodes in a model}
\usage{
get_parents(model)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by
\code{\link{make_model}}.}
}
\value{
A \code{list} of parents in a DAG
}
\description{
Get list of parents of all nodes in a model
}
