% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_models.R
\name{get_parents}
\alias{get_parents}
\title{Get list of parents in a dag}
\usage{
get_parents(model)
}
\arguments{
\item{A}{probabilistic causal model created by make_model()}
}
\value{
A list of parents in a DAG
}
\description{
Get list of parents in a dag
}
\examples{

\dontrun{
dag <- make_dag(
 add_edges(parent = "X",children = c("K","Y")),
 add_edges(parent = "K",children = "Y")
)

get_parents(dag)
}

dag <- make_dag(
 add_edges(parent = "X",children = c("K")),
 add_edges(parent = "K",children = "Y")
)

get_parents(dag)
get_ancestors(dag)
}
