% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmap.R
\name{get_parmap}
\alias{get_parmap}
\title{Get parmap: a matrix mapping from parameters to data types}
\usage{
get_parmap(model, A = NULL, P = NULL)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by
\code{\link{make_model}}.}

\item{A}{A \code{data.frame}. Ambiguities matrix. Not required but may be
provided to avoid repeated computation for simulations. \code{inspect(model, "ambiguities_matrix")}}

\item{P}{A \code{data.frame}. Parameter matrix. Not required but may be
provided to avoid repeated computation for simulations. See \code{inspect(model, "parameter_matrix")}.}
}
\value{
A matrix
}
\description{
Gets parmap from a model, or generates if not available.
}
