% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_possible_data.R
\name{get_possible_data}
\alias{get_possible_data}
\title{Get Possible Data}
\usage{
get_possible_data(model, collapse = TRUE)
}
\arguments{
\item{model}{A probabilistic causal model created by \code{make_model()}}

\item{collapse}{whether to collapse possible data}
}
\description{
Creates a list of all possible data realizations for each node given each possible parent value
}
