% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan.R
\name{get_possible_data_internal}
\alias{get_possible_data_internal}
\title{get_possible_data_internal}
\usage{
get_possible_data_internal(model, collapse = TRUE)
}
\arguments{
\item{model}{A model created by \code{make_model}}
}
\value{
a list
}
\description{
get_possible_data_internal
}
\keyword{internal}
