% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_priors}
\alias{get_priors}
\title{Get prior distribution}
\usage{
get_priors(model, priors = NULL, prior_distribution = "uniform",
  alphas = NULL)
}
\arguments{
\item{model}{A model object generated by make_model().}

\item{priors}{A numeric vector. Dirichlet hyperparameters.}

\item{prior_distribution}{A character vector. Indicates the prior distribution to be used. Can be 'uniform', 'jeffreys', or 'certainty'.}

\item{alphas}{A numeric vector. Hyperparameters of the Dirichlet distribution to be passed to \code{make_priors()}.}
}
\description{
Get prior distribution
}
