% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_priors.R
\name{get_priors}
\alias{get_priors}
\title{Get priors}
\usage{
get_priors(model)
}
\arguments{
\item{model}{A model object generated by make_model().}
}
\description{
Extracts priors as a named vector
}
\examples{
get_priors(make_model("X -> Y"))
}
