% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query_types.R
\name{get_query_types}
\alias{get_query_types}
\title{Look up query types}
\usage{
get_query_types(model, query, map = "causal_type")
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}

\item{query}{A character string. An expression defining nodal types to interrogate \code{\link{reveal_outcomes}}}

\item{map}{Types in query. Either \code{nodal_type} or \code{causal_type}. Default is causal types}
}
\description{
Find the nodal or causal types used by the query. Passes arguments into \code{\link{map_query_to_causal_type}} or \code{\link{map_query_to_nodal_type}}
}
\examples{
model <- make_model('X -> M -> Y; X->Y')
query <- '(Y[X=0] > Y[X=1])'
get_query_types(model, query, map="nodal_type")
get_query_types(model, query, map="causal_type")
get_query_types(model, query)
}
