% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query_types.R
\name{get_query_types}
\alias{get_query_types}
\title{Get values of types according to a query}
\usage{
get_query_types(model, query, join_by = "|")
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}

\item{query}{A character string. An expression defining nodal types to interrogate \code{\link{reveal_outcomes}}}

\item{join_by}{A logical operator. Used to connect causal statements: \emph{AND} ('&') or \emph{OR} ('|'). Defaults to '|'.}
}
\value{
A \code{list} containing the types and the evaluated expression.
}
\description{
Get values of types according to a query
}
\examples{
model <- make_model('X -> M -> Y; X->Y')
query <- '(Y[X=1] > Y[X=0]) & (M[X=0]==1)'
x <- get_query_types(model, query)
summary(x)

query <- 'Y[M=M[X=0], X=1]==1'
x <- get_query_types(model, query)
get_query_types(model, query)

query <- '(Y[X=1, M = 1] >  Y[X=0, M = 1]) & (Y[X=1, M = 0] >  Y[X=0, M = 0])'
get_query_types(model, query)

query <- 'Y[X=1] == Y[X=0]'
get_query_types(model, query)

query <- '(X == 1) & (M==1) & (Y ==1) & (Y[X=0] ==1)'
x <- get_query_types(model, query)

query <- '(Y[X = .]==1)'
get_query_types(model, query)

model <- make_model('X -> Y')
}
