% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_dags.R
\name{get_terminal_vars}
\alias{get_terminal_vars}
\title{Get endogenous variables in a DAG which are also terminal}
\usage{
get_terminal_vars(pcm)
}
\arguments{
\item{dag}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
