% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_restrictions.R
\name{get_type_names}
\alias{get_type_names}
\title{Get type names}
\usage{
get_type_names(nodal_types)
}
\arguments{
\item{nodal_types}{Nodal types of a model. See \code{\link{get_nodal_types}}.}
}
\value{
A vector containing causal type names
}
\description{
Get type names
}
\examples{
model <- make_model('A->Y<-B')
get_type_names(model$nodal_types)
}
\keyword{internal}
