% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_dags.R
\name{get_types}
\alias{get_types}
\title{Get list of types for variables in a DAG}
\usage{
get_types(dag)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
A list of parents in a DAG
}
\description{
Get list of types for variables in a DAG
}
\examples{

dag <- make_dag(
 add_edges(parent = "X",children = c("K","Y")),
 add_edges(parent = "K",children = "Y")
)

get_types(dag)


}
