% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_models.R
\name{get_types}
\alias{get_types}
\title{Get values of types according to a query}
\usage{
get_types(model, query)
}
\arguments{
\item{model}{A model  created by \code{make_model}}

\item{query}{A quoted expression to interrogate \code{reveal_outcomes}}
}
\value{
A list containing the types and the evaluated expression
}
\description{
Get values of types according to a query
}
\examples{
model <- make_model("X -> M -> Y; X -> Y")
query <- "(Y[X=1] > Y[X=0]) & (M[X=0]==1)"
x <- get_types(model, query)
x$types[x$types]
query <- "Y[M=M[X=0], X=1]==1"
x <- get_types(model, query)
x$types[x$types]
}
