% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_types.R
\name{get_types}
\alias{get_types}
\title{Get values of types according to a query}
\usage{
get_types(model, query)
}
\arguments{
\item{model}{A model  created by \code{make_model}}

\item{query}{A quoted expression to interrogate \code{reveal_outcomes}}
}
\value{
A list containing the types and the evaluated expression
}
\description{
Get values of types according to a query
}
\examples{
model <- make_model("X"\%->\%"M", "X"\%->\%"Y", "M"\%->\%"Y")
query <- "(Y[X=1] > Y[X=0]) & (M[X=0]==1)"
x <- get_types(model, query)
x$types[x$types]
query <- "Y[M=M[X=0], X=1]==1"
x <- get_types(model, query)
get_types(model, query)
query <- "(Y[X=1, M = 1] >  Y[X=0, M = 1]) & (Y[X=1, M = 0] >  Y[X=0, M = 0])"
get_types(model, query)
query <- "Y[X=1] == Y[X=0]"
get_types(model, query)
}
