% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspecting_dags.R
\name{get_variables}
\alias{get_variables}
\title{Get variable names from a DAG}
\usage{
get_variables(pcm)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
A vector of variable names
}
\description{
Get variable names from a DAG
}
