% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{get_weights_ambiguities}
\alias{get_weights_ambiguities}
\title{get_weights_ambiguities}
\usage{
get_weights_ambiguities(model, data)
}
\arguments{
\item{model}{A model generated by \code{make_model()}}

\item{data}{A data.frame of variables that can take three values: 0, 1, and NA.}
}
\value{
Returns indices and ambiguity matrix
}
\description{
get_weights_ambiguities
}
