% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab}
\alias{grab}
\title{Grab}
\usage{
grab(model, object = NULL, ...)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by
\code{\link{make_model}}.}

\item{object}{A character string specifying the component to retrieve.
Available options are:
\itemize{
\item \code{"dag"} A data frame with columns ‘parent’ and ‘children’ indicating how nodes relate to each other,
\item \code{"statement"} a character. Statement describing causal relations using dagitty syntax,
\item \code{"nodes"} A list containing the nodes in the model,
\item \code{"parents_df"} a table listing nodes, whether they are root nodes or not, and the number of parents they have,
\item \code{"parameters_df"} a data frame containing parameter information,
\item \code{"nodal_types"} a list with the nodal types of the model,
\item \code{"causal_types"} a data frame listing causal types and the nodal types that produce them,
\item \code{"parameter_matrix"} a matrix mapping from parameters into causal types,
\item \code{"ambiguities_matrix"} a matrix mapping from causal types into data types,
\item \code{"parameter_prior"} named vector of the parameter prior,
\item \code{"priors"}  a data frame of the parameter prior distribution,
\item \code{"posteriors"}  a data frame of the parameter posterior distribution,
\item \code{"parameters"} a vector of 'true' parameters,
\item \code{"stan_object"}  stan_objects can include the stanfit object and distributions over nodal types and event probabilities (w).
\item \code{"stan_fit"} the stanfit object generated by Stan.
\item \code{"stan_fit_summary"} a summary of the stanfit object generated by Stan.
\item \code{"type_posterior"} a matrix of type probabilities using posteriors,
\item \code{"type_prior"} a matrix of type probabilities using priors,
\item \code{"data"} the data used for updating the model,
\item \code{"event_probabilities"} a matrix mapping from parameters to event probabilities,
}}

\item{...}{Other arguments passed to helper \code{get_*} functions.}
}
\value{
Objects from a \code{causal_model} as specified.
}
\description{
Returns specified elements from a \code{causal_model}.
Users can use \code{grab} to extract model's components including
nodal types, causal types, parameter priors, parameter posteriors,
type priors, type posteriors, and other relevant elements.
See argument \code{object} for other
}
\examples{
\donttest{
model <- make_model('X -> K -> Y')
grab(model, object = "dag")
grab(model, object = "parameters_df")
grab(model, object = "nodal_types")
grab(model, object = "causal_types")

model <- make_model('X -> Y') |> update_model()
grab(model, object = "stan_fit_summary")

}

}
