% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{includes_var}
\alias{includes_var}
\title{Whether a query contains an exact string}
\usage{
includes_var(var, query)
}
\arguments{
\item{var}{Variable name}

\item{query}{An expression in string format.}
}
\value{
A logical expression indicating whether a variable is included in a query
}
\description{
Whether a query contains an exact string
}
\keyword{Used}
\keyword{in}
\keyword{internal}
\keyword{map_query_to_nodal_types}
