% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_helpers.R
\name{interacts}
\alias{interacts}
\title{Make statement for any interaction}
\usage{
interacts(X1, X2, Y)
}
\arguments{
\item{X1}{input node 1}

\item{X2}{input node 2}

\item{Y}{outcome node}
}
\value{
A character statement of class statement
}
\description{
Generate a statement for X1, X1 interact in the production of Y
}
\examples{
interacts("A", "B", "Y")
get_query_types(model = make_model("X-> Y <- W"),
         query = interacts("X", "W", "Y"))

}
\seealso{
Other statements: \code{\link{complements}},
  \code{\link{decreasing}}, \code{\link{increasing}},
  \code{\link{substitutes}}
}
\concept{statements}
