% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type_helpers.R
\name{list_non_parents}
\alias{list_non_parents}
\title{Returns a list with the nodes that are not directly pointing into a node}
\usage{
list_non_parents(model, node)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}

\item{node}{A character string. The quoted name of a node.}
}
\description{
Returns a list with the nodes that are not directly pointing into a node
}
\keyword{internal}
