% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lookup_type}
\alias{lookup_type}
\title{Interpret or find position}
\usage{
lookup_type(model, condition = NULL, position = NULL)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{condition}{a condition}

\item{position}{a position}
}
\description{
Interprets the causal type of a given position in node, or returns positions that correspond to given condition.
}
