% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_alphas}
\alias{make_alphas}
\title{Make alphas}
\usage{
make_alphas(model, alphas = NULL)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{alphas}{A numeric vector. Hyperparameters of the Dirichlet distribution to be passed to \code{make_priors()}.}
}
\description{
Stipulated alpha values override prior_distribution
}
