% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_confounds_df.R
\name{make_confounds_df}
\alias{make_confounds_df}
\title{Get confounds dataframe}
\usage{
make_confounds_df(model, ...)
}
\arguments{
\item{model}{A model made by make_model}

\item{...}{arguments passed to get_nodal_joint_probability}
}
\description{
Identifies confounded nodal types. Default behavior draws a random admissible parameter vector.
}
\examples{
model <- make_model("X -> M -> Y; X <-> M")
make_confounds_df(model)

model <- make_model("X -> M -> Y") \%>\%
  set_confound(list(X = "Y[X=1] > Y[X=0]"))
make_confounds_df(model)
}
