% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_dags.R
\name{make_dag}
\alias{make_dag}
\title{Make a DAG}
\usage{
make_dag(...)
}
\arguments{
\item{...}{Edges added by add_edges()}
}
\value{
An object of class probabilistic_causal_model containing a dag
}
\description{
Make a DAG
}
\examples{
\dontrun{
dagXKY <- make_dag(
add_edges(parent = "X",children = c("K","Y")),
add_edges(parent = "K",children = "Y")
)
}
}
