% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_strategies.R
\name{make_data_probabilities}
\alias{make_data_probabilities}
\title{Generates a probability distribution over possible data outcomes}
\usage{
make_data_probabilities(model, pars, possible_data)
}
\arguments{
\item{model}{A causal model as created by \code{make_model}}

\item{given}{A data frame with observations}

\item{subset}{data strategy}
}
\value{
A dataset
}
\description{
NOTE: This needs to be checked for whether it is taking account of strategy probabilities properly
}
\examples{

library(dplyr)
model <- make_model("X->M->Y")  \%>\%
   set_restrictions(causal_type_restrict = "Y[M=1]<Y[M=0] | M[X=1]<M[X=0] ") \%>\%
   set_parameter_matrix()

pars <- draw_parameters(model)
possible_data <- make_possible_data(model, N= 2)
make_data_probabilities(model, pars = pars, possible_data)

given <- data.frame(X = c(0,0,0,1,1,1), M = NA, Y = c(0,0,1,0,1,1))
possible_data <- make_possible_data(model, given = given, cases = "X==1 & Y==1")
make_data_probabilities(model, pars = pars, possible_data)

}
