% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan.R
\name{make_gbiqq_data}
\alias{make_gbiqq_data}
\title{Prepare data for stan}
\usage{
make_gbiqq_data(model, data, P = NULL)
}
\arguments{
\item{model}{A probabilistic causal model created by \code{make_model}}

\item{data}{A data frame with observations}

\item{P}{A matrix mapping parameters (rows) to types (columns). If not provided, defaults one parameter per nodal type.}

\item{lambdas_prior}{A vector containg priors for lambda}
}
\value{
a list
}
\description{
Create a list containing the data to be passed to stan
}
\keyword{internal}
