% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan.R
\name{make_gbiqq_data}
\alias{make_gbiqq_data}
\title{Prepare data for stan}
\usage{
make_gbiqq_data(model, data)
}
\arguments{
\item{model}{A model created by \code{make_model}}

\item{data}{A data frame with observations}

\item{lambdas_prior}{A vector containg priors for lambda}

\item{P}{A matrix mapping parameters (rows) to types (columns). If not provided, defaults one parameter per nodal type.}
}
\value{
a list
}
\description{
Create a list containing the data to be passed to stan
}
\examples{
model <- make_model(add_edges(parent = "X", children = "Y"))
data   <- draw_data(model, n = 6)
make_gbiqq_data(model, data)
data[1,1]   <- NA
make_gbiqq_data(model, data)


}
