% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepping_stan_using_dags.R
\name{make_gbiqq_data}
\alias{make_gbiqq_data}
\title{Prepare data for stan}
\usage{
make_gbiqq_data(dag, data, lambdas_prior)
}
\arguments{
\item{dag}{A dag as created by \code{make_dag}}

\item{data}{A data frame with observations}

\item{lambdas_prior}{A vector containg priors for lambda}
}
\value{
a list
}
\description{
Create a list containing the data to be passed to stan
}
\keyword{internal}
