% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gbiqq_data.R
\name{make_gbiqq_data}
\alias{make_gbiqq_data}
\title{Prepare data for stan}
\usage{
make_gbiqq_data(model, data)
}
\arguments{
\item{model}{A model created by \code{make_model}}

\item{data}{A data frame with observations}
}
\value{
a list
}
\description{
Create a list containing the data to be passed to stan
}
\examples{
model <- make_model("X->Y")
data <- simulate_data(model, n = 6)
make_gbiqq_data(model, data)
data[1,1]  <- NA
make_gbiqq_data(model, data)


}
