% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{make_model}
\alias{make_model}
\title{Make a model}
\usage{
make_model(statement, add_causal_types = TRUE)
}
\arguments{
\item{statement}{A character. Statement describing causal
relations using dagitty syntax. Only directed relations are permitted.
For instance "X -> Y" or  "X1 -> Y <- X2; X1 -> X2".}

\item{add_causal_types}{Logical. Whether to dreate and attach causal types to \code{model}. Defaults to `TRUE`.}
}
\value{
An object of class \code{causal_model} containing a DAG.
}
\description{
\code{make_model} uses \link{dagitty} syntax and functionality to specify nodes and edges of a
graph. Implied causal types are calculated and default priors are provided under the
assumption of no confounding.
Models can be updated with specification of a parameter matrix, \code{P}, by
providing restrictions on causal types, and/or by providing informative priors on parameters.
The default setting for a causal model have flat (uniform) priors and parameters
putting equal weight on each parameter within each parameter set. These can be
adjust with \code{set_priors} and \code{set_parameters}
}
\examples{
make_model(statement = "X -> Y")
modelXKY <- make_model("X -> K -> Y; X -> Y")

# Example where cyclicaly dag attempted
\dontrun{
 modelXKX <- make_model("X -> K -> X")
}

# Examples with confounding
model <- make_model("X->Y; X <-> Y")
model$P
model <- make_model("Y2 <- X -> Y1; X <-> Y1; X <-> Y2")
model$P
model$confounds_df
dim(model$P)
model$P
model <- make_model("X1 -> Y <- X2; X1 <-> Y; X2 <-> Y")
dim(model$P)
model$parameters_df

# A single node graph is also possible
model <- make_model("X")
plot(model)

# Unconnected nodes cannot
\dontrun{
 model <- make_model("X <-> Y")
 plot(model)
}
}
