% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_models.R
\name{make_model}
\alias{make_model}
\title{Make a model}
\usage{
make_model(..., add_priors = FALSE)
}
\arguments{
\item{...}{Edges added by add_edges()}
}
\value{
An object of class probabilistic_causal_model containing a dag
}
\description{
`make_model` requires specifying nodes and edges of a graph. Imlpied causal types are calculated and default priors are provided under the assumption of no confounding.
Models can be updated with specification of a parameter matrix, P, by providing restrictions on causal types, and/or by providing informative priors on parameters. `
}
\examples{
\dontrun{
modelXKY <- make_model(
add_edges(parent = "X",children = c("K","Y")),
add_edges(parent = "K",children = "Y")
)

# Example where variables need re-ordering
model <- make_model(
  add_edges(parent = "X3",  children = "Y"),
  add_edges(parent = "X2", children = "X3"),
  add_edges(parent = "X2", children = "Y"),
  add_edges(parent = "X1", children = "X2"))

# Example where cyclicaly dag attempted
model <- make_model(
  add_edges(parent = "X1",  children = "X2"),
  add_edges(parent = "X2", children = "X3"),
  add_edges(parent = "X3", children = "X2"))
}
}
