% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_nodal_types}
\alias{make_nodal_types}
\title{Make nodes from matrix}
\usage{
make_nodal_types(mat)
}
\arguments{
\item{mat}{A character matrix. For each variable contained in a model, matrix contains named columns with nodal values.}
}
\value{
A named list of causal types for each variable
}
\description{
Make nodes from matrix
}
