% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodal_types.R
\name{make_nodal_types}
\alias{make_nodal_types}
\title{Make nodal types}
\usage{
make_nodal_types(model)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}
}
\description{
Make nodal types
}
