% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodal_types.R
\name{make_nodal_types}
\alias{make_nodal_types}
\title{Make nodal types}
\usage{
make_nodal_types(model, include_node_names = FALSE)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}

\item{include_node_names}{Logical. If `TRUE` returns names of form X0, X1; otherwise returns 0, 1. Defaults to `FALSE`}
}
\value{
A named list containing nodal types for each node
}
\description{
Make nodal types
}
\examples{
\donttest{
model <- make_model('X -> K -> Y')
CausalQueries:::make_nodal_types(model)
}
}
\keyword{internal}
