% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_matrix.R
\name{make_parameter_matrix}
\alias{make_parameter_matrix}
\title{Make parameter matrix}
\usage{
make_parameter_matrix(model)
}
\arguments{
\item{model}{A model created by make_model()}

\item{confound}{A list relating nodes to types with whihc  they are confounded}
}
\description{
Calculate parameter matrix assuming no confounding. The parameter matrix  maps from parameters into causal types. In models without confounding parameters correspond to nodal types.
}
\examples{
model <- make_model("X -> Y")
make_parameter_matrix(model)
}
