% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmap.R
\name{make_parmap}
\alias{make_parmap}
\title{Make parmap: a matrix mapping from parameters to data types}
\usage{
make_parmap(model, A = NULL, P = NULL)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by
\code{\link{make_model}}.}

\item{A}{A \code{data.frame}. Ambiguity matrix. Not required but may be
provided to avoid repeated computation for simulations.}

\item{P}{A \code{data.frame}. Parameter matrix. Not required but may be
provided to avoid repeated computation for simulations.}
}
\value{
A matrix
}
\description{
Generates a matrix with a row per parameter and a column per data type.
}
