% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_strategies.R
\name{make_possible_data_single}
\alias{make_possible_data_single}
\title{Make possible data for a single strategy}
\usage{
make_possible_data_single(model, given, data_strat = NULL,
  vars_to_seek = NA)
}
\description{
Make possible data for a single strategy
}
