% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_priors}
\alias{make_priors}
\title{Set priors}
\usage{
make_priors(model, prior_distribution = "uniform", alphas = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{prior_distribution}{A character indicating the prior distribuiton}

\item{alphas}{the hyperparameters of the dirichlet distribution.}
}
\description{
Set priors
}
\examples{
XYdag <- make_model("X -> Y")
#  Default sets all priors to 1
make_priors(model = XYdag)
#  Set all priors to 0.5
make_priors(model = XYdag, prior_distribution = "jeffreys")
#  Set all priors to infinity
make_priors(model = XYdag, prior_distribution = "certainty")

#  set all priors to 1 except for prior of nodal_type X0
make_priors(model = XYdag, alpha = list(X = c(X0 = 2)))
#  specify priors for each of the nodal_types in model
make_priors(model = XYdag, prior_distribution = "uniform" ,
           alpha = list(X = c(X0 = 2, X1 = 1),  Y = c(Y00 = 1, Y01 = 2, Y10 = 2, Y11 = 1)))
# set all priors to 10
make_priors(model = XYdag, prior_distribution = NULL,
           alpha =  10)
}
