% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mak_priors2_v4.R
\name{make_priors2}
\alias{make_priors2}
\title{make_priors2}
\usage{
make_priors2(model, prior_distribution = NA, parameter_set = NA,
  statement = NA, confound = NA, alphas = NULL, label = NA)
}
\arguments{
\item{model}{A mode created with make_model}

\item{prior_distribution}{A common priors distribution (uniform, jeffreys or certainty)}

\item{parameter_set}{A node}

\item{statement}{A query}

\item{confound}{A confound statement}

\item{alphas}{Hyperparamenters of the Dirichlet distribution}

\item{label}{The name of a nodal type}
}
\description{
This function creates the priors to be passed on nodal types
}
\details{
Function takes as arguments the following:
}
