% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_priors_single}
\alias{make_priors_single}
\title{make_priors_single}
\usage{
make_priors_single(model, prior_distribution = NA, alphas = NULL,
  node = NA, statement = NA, confound = NA, label = NA)
}
\arguments{
\item{model}{A mode created with \code{make_model}}

\item{prior_distribution}{A common prior distribution (uniform, jeffreys or certainty)}

\item{alphas}{Hyperparameters of the Dirichlet distribution}

\item{node}{A node}

\item{statement}{A query}

\item{confound}{A confound statement}

\item{label}{The name of a nodal type}
}
\description{
This is the one step function for make_priors, it creates the priors to be passed on nodal types with \code{set_priors}. See \code{make_priors} for more help.
}
\details{
Forbidden statements include:
\itemize{
  \item Setting \code{prior_distribution} and \code{alphas} at the same time.
  \item Setting a \code{prior_distribution} other than uniform, jeffreys or certainty.
  \item Setting negative priors.
}
}
\examples{
model <- make_model("X -> M -> Y; X->Y")

make_priors_single(model, prior_distribution = "jeffreys")
make_priors_single(model, alphas = 1:2)

make_priors_single(model, statement = "(Y[X=0, M = .] > Y[X=1, M = 0])", alphas = 666)
make_priors_single(model, parameter_set = "Y", statement = "(Y[X=0, M = .] > Y[X=1, M = 0])", alphas = 777)

make_priors_single(model, parameter_set = "M", alphas = 888)

make_priors_single(model, label = "X0", alphas = 999)

model <- make_model("X->Y")
model <- set_confound(model, list(X = "Y[X=1] > Y[X=0]"))
model <- set_confound(model, list(X = "Y[X=1] < Y[X=0]"))
make_priors_single(model, confound = "Y[X=1] > Y[X=0]", alphas = 3)
make_priors_single(model, parameter_set = "X", confound = "Y[X=1] > Y[X=0]", alphas = 3)
}
