% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{minimal_data}
\alias{minimal_data}
\title{Creates a data frame for case with no data}
\usage{
minimal_data(model)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}
}
\value{
A \code{data.frame} with one row of NAs and columns named according to nodes in a model.
}
\description{
Creates a data frame for case with no data
}
\examples{
\donttest{
model <- make_model('X -> K -> Y')
CausalQueries:::minimal_data(model)
}
}
\keyword{internal}
