% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{minimal_event_data}
\alias{minimal_event_data}
\title{Creates a compact data frame for case with no data}
\usage{
minimal_event_data(model)
}
\description{
FLAG: In principle this might produce conflicts with node reduction
}
