% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{minimal_event_data}
\alias{minimal_event_data}
\title{Creates a compact data frame for case with no data}
\usage{
minimal_event_data(model)
}
\arguments{
\item{model}{A causal model made by make_model
FLAG: In principle this might produce conflicts with node reduction}
}
\description{
Creates a compact data frame for case with no data
}
