% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type_helpers.R
\name{nodes_in_statement}
\alias{nodes_in_statement}
\title{Identify nodes in a statement}
\usage{
nodes_in_statement(nodes, statement)
}
\arguments{
\item{nodes}{A vector of characters. It should contain quoted names of
the nodes in \code{model}}

\item{statement}{A character. A quoted causal statement.}
}
\value{
Returns name of nodes present in a statement
}
\description{
Identify nodes in a statement
}
\keyword{internal}
