% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realise_outcomes.R
\name{parents_to_int}
\alias{parents_to_int}
\title{Helper to turn parents_list into a list of data_realizations column positions}
\usage{
parents_to_int(parents_list, position_set)
}
\arguments{
\item{parents_list}{a named list of character vectors specifying all nodes in the DAG and their respective parents}
}
\value{
a list of column positions
}
\description{
Helper to turn parents_list into a list of data_realizations column positions
}
\keyword{Used}
\keyword{in}
\keyword{internal}
\keyword{realise_outcomes}
