% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{perm}
\alias{perm}
\title{Produces the possible permutations of a set of nodes}
\usage{
perm(max = rep(1, 2))
}
\arguments{
\item{max}{A vector of integers. The maximum value of an integer value
starting at 0. Defaults to 1. The number of permutation is defined
by \code{max}'s length}
}
\value{
A \code{matrix} of permutations
}
\description{
Produces the possible permutations of a set of nodes
}
\keyword{internal}
