% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{plot_dag}
\alias{plot_dag}
\title{Plot your dag using 'dagitty'}
\usage{
plot_dag(model)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}
}
\value{
No return value.
}
\description{
Plot your dag using 'dagitty'
}
\examples{
\donttest{
model <- make_model('X -> K -> Y; X -> Y')
CausalQueries:::plot_dag(model)
model <- CausalQueries:::make_model('X -> K -> Y; X <-> Y')
CausalQueries:::plot_dag(model)
}

}
\keyword{internal}
