% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{plot_dag}
\alias{plot_dag}
\title{Plots a DAG in ggplot style using a causal model input}
\usage{
plot_dag(
  model = NULL,
  x = NULL,
  y = NULL,
  title = "",
  textcol = "white",
  textsize = 3.88,
  obscure = NULL,
  shape = 16,
  nodecol = "black",
  nodesize = 16
)
}
\arguments{
\item{model}{A \code{causal_model} object generated from \code{make_model}}

\item{x}{A vector of coordinates for DAG nodes. If left empty, coordinates are randomly generated}

\item{y}{A vector of coordinates for DAG nodes. If left empty, coordinates are randomly generated}

\item{title}{String specifying title of graph}

\item{textcol}{String specifying colour of text labels}

\item{textsize}{Numeric, size of text labels}

\item{obscure}{A vector or string of the form "X->Y" that indicates if a given arrow should be hidden}

\item{shape}{Indicates shape of node. Defaults to circular node.}

\item{nodesize}{Size of node.}

\item{nodelcol}{String indicating colour of node that is accepted by ggplot's default palette}
}
\value{
A DAG plot in ggplot style.
}
\description{
If confounds are indicated (provided in \code{attr(model$P, 'confounds')}), then these are represented as bidirectional arcs.
}
\examples{

\dontrun{
model <- CausalQueries:::make_model('X -> K -> Y; X <-> Y')
CausalQueries:::plot_dag(model)
}

}
\keyword{internal}
