% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representing_dags.R
\name{plot_dag}
\alias{plot_dag}
\title{Plot your dag using dagitty}
\usage{
plot_dag(dag)
}
\arguments{
\item{dag}{A dag created by make_dag()}
}
\value{
dagitty translation of DAG
}
\description{
Plot your dag using dagitty
}
\examples{
\dontrun{
dag <- make_dag(
 add_edges(parent = "X",children = c("K","Y")),
 add_edges(parent = "K",children = "Y")
)

plot_dag(dag)
}

}
