% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{plot_dag}
\alias{plot_dag}
\title{Plot your dag using dagitty}
\usage{
plot_dag(model)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}
}
\value{
dagitty translation of DAG
}
\description{
Plot your dag using dagitty
}
\examples{
\dontrun{
model <- make_model('X -> K -> Y; X -> Y')
plot_dag(model)
model <- make_model('X -> K -> Y; X <-> Y')
plot_dag(model)
}

}
