% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{position_nodes}
\alias{position_nodes}
\title{internal helper
position_nodes generates a force directed graph layout to optimally
position nodes. Connected nodes attract while unconnected nodes repel.
positions_nodes additionally attempts to minimize edge intersections
and maximize layout compactness.}
\usage{
position_nodes(
  dag,
  nodes,
  iterations = 1000L,
  k = 1,
  repulsion = 0.1,
  desired_distance = 0.5,
  crossing_penalty = 0.1,
  gravity = 0.1
)
}
\description{
internal helper
position_nodes generates a force directed graph layout to optimally
position nodes. Connected nodes attract while unconnected nodes repel.
positions_nodes additionally attempts to minimize edge intersections
and maximize layout compactness.
}
\keyword{internal}
