% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/query_model.R
\name{print.model_query}
\alias{print.model_query}
\alias{plot.model_query}
\title{Print a tightened summary of model queries}
\usage{
\method{print}{model_query}(x, ...)

\method{plot}{model_query}(x, ...)
}
\arguments{
\item{x}{an object of class \code{model_query}}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{model_query}.
}
