% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_model.R
\name{print.model_query}
\alias{print.model_query}
\title{Print a tightened summary of model queries}
\usage{
\method{print}{model_query}(x, digits = 3)
}
\arguments{
\item{x}{An object of \code{model_query} class.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
print method for class \code{model_query}.
}
