% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_model.R
\name{queries_to_types}
\alias{queries_to_types}
\title{helper to get types from queries}
\usage{
queries_to_types(jobs, model, query_col, realisations)
}
\arguments{
\item{jobs}{a data frame of argument combinations}

\item{model}{a list of models}

\item{query_col}{string specifying the name of the column in jobs
holding queries to be evaluated}

\item{realisations}{list of data frame outputs from calls
to \code{realise_outcomes}}
}
\value{
jobs data frame with a nested column of
  \code{map_query_to_nodal_type} outputs
}
\description{
helper to get types from queries
}
\keyword{internal}
