% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_and_estimands.R
\name{query_distribution}
\alias{query_distribution}
\title{Calculate estimand distribution}
\usage{
query_distribution(model, query, subset = TRUE, using = "priors",
  parameters = NULL, type_distribution = NULL, verbose = FALSE,
  join_by = "|")
}
\arguments{
\item{model}{A model}

\item{query}{A query on potential outcomes such as "Y[X=1] - Y[X=0]"}

\item{subset}{quoted expression evaluates to logical statement. subset allows estimand to be conditioned on *observational* distribution.}

\item{using}{String indicating whether to use `priors`, `posteriors` or `parameters`}

\item{parameters}{A true parameter vector to be used instead of parameters attached to the model in case  `using` specifies `parameters`}

\item{type_distribution}{if provided saves calculation, otherwise clculated from model; may be based on prior or posterior}

\item{verbose}{Logical. Whether to print mean and standard deviation of the estimand on the consule.}

\item{join_by}{A string. The logical operator joining expanded types when \code{causal_type_restrict} contains wildcard (\code{.}). Can take values \code{"&"} (logical AND) or \code{"|"} (logical OR). When restriction contains wildcard (\code{.}) and \code{join_by} is not specified, it defaults to \code{"|"}, otherwise it defaults to \code{NULL}.}
}
\description{
Calculated from a prior or posterior distribution
}
\examples{
library(dplyr)
model <- make_model("X -> Y") \%>\%
         set_prior_distribution()
 query_distribution(model, query = "(Y[X=1] - Y[X=0])")
 query_distribution(model, query = "(Y[X=1] - Y[X=0])", subset = "X==1")
 query_distribution(model, query = "(Y[X=1] - Y[X=0])", subset = "Y[X=1]==1")
 query_distribution(model, query = "(Y[X=1] > Y[X=0])")
 query_distribution(model, query = "(Y[X=.] == 1)", join_by = "&")
 query_distribution(model, query = "(Y[X=1] - Y[X=0])", using = "posteriors")
 query_distribution(model, query = "(Y[X=1] - Y[X=0])", using = "parameters")
}
