% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_and_estimands.R
\name{query_model}
\alias{query_model}
\title{Generate estimands dataframe}
\usage{
query_model(model, parameters = NULL, queries = list(NULL),
  subsets = list(TRUE), using = list("priors"), stats = NULL,
  digits = 3, n_draws = 4000, expand_grid = FALSE)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{parameters}{A numeric vector of  true parameters, if available.}

\item{queries}{A string. Query on potential otcomes such as "Y[X=1] - Y[X=0]".}

\item{subsets}{A quoted expression that evaluates to a logical statement. Allows estimand to be conditioned on *observational* distribution.}

\item{using}{String indicating whether to use `priors`, `posteriors` or `parameters`.}

\item{stats}{Functions be applied to estimand distribution. If `NULL`, defaults to mean and standard deviation.}

\item{digits}{An integer. Decimal digits in output table.}

\item{n_draws}{An integer. Number of draws.}

\item{expand_grid}{logical If TRUE then all combinations of provided lists are examined. If not then each list is cycled through separately.}
}
\description{
Calculated from a parameter vector, from a prior or from a posterior distribution
}
\examples{
library(dplyr)
model <- make_model("X -> Y") \%>\% set_prior_distribution(n_draws = 10000)

estimands_df <-query_model(
               model,
               queries = list(ATE = "Y[X=1] - Y[X=0]", Share_positive = "Y[X=1] > Y[X=0]"),
               using = c("parameters", "priors"),
               expand_grid = TRUE)

estimands_df <-query_model(
               model,
               queries = list(ATE = "Y[X=1] - Y[X=0]", Share_positive = "Y[X=1] > Y[X=0]"),
               using = c("parameters", "priors"),
               expand_grid = FALSE)

estimands_df <- query_model(
                model,
                using = list( "parameters", "priors"),
                queries = list(ATE = "Y[X=1] - Y[X=0]", Is_B = "Y[X=1] > Y[X=0]"),
                subsets = list(TRUE, "Y==1 & X==1", "Y==0 & X==1"),
                digits = 3)

}
