% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_and_estimands.R
\name{query_model}
\alias{query_model}
\title{Query a model}
\usage{
query_model(model, query, subset = TRUE, w = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{query}{A logic operation on variables, in quotes. For example "Y==1"}

\item{subset}{An optional logic condition on variabless, in quotes.. For example "X==0"}
}
\description{
Ask a question of a model
}
\examples{
model <- make_model(add_edges(parent = "X", children = c("Y")))
w     <- draw_event_prob(model, P = NULL, A = NULL)
query_model(model,  "Y==1", "X==1", w)
query_model(model,  "Y==1", "X==1", w) - query_model(model,  "Y==1", "X==0", w)

# ATE: Make a model with monotonicity and reasonable prior certainty on uniform types
model     <-  make_model(add_edges(parent = "X", children = c("Y")))
model     <-  set_priors(model = model, prior_distribution = NULL, alpha =  10)
model     <-  reduce_nodal_types(model = model, restrictions = list(Y = "Y10"))
model_do0 <-  reduce_nodal_types(model = model, do = list(X = 0))
model_do1 <-  reduce_nodal_types(model = model, do = list(X = 1))
query_model(model_do1,  "Y==1") - query_model(model_do0,  "Y==1")
simulations <- replicate(200, query_model(model_do1,  "Y==1") - query_model(model_do0,  "Y==1"))
hist(simulations, main = paste("E(ATE) = ", round(mean(simulations), 2)))
# This is not quite right since the *same* prior draw should be used for the two queries -- a bit confusing though
# as currently set up since the two models have different parameters! May need another approch to "do":
# e.g. draw a parameter vector and set certainty around that, calculated estimand, repeat over different draws
}
