% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_type.R
\name{query_to_expression}
\alias{query_to_expression}
\title{Helper to turn query into a data expression}
\usage{
query_to_expression(query, node)
}
\description{
Helper to turn query into a data expression
}
