% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_type.R
\name{query_to_expression}
\alias{query_to_expression}
\title{Helper to turn query into a data expression}
\usage{
query_to_expression(query, node)
}
\arguments{
\item{query}{A character string. An expression defining nodal types to interrogate \code{\link{reveal_outcomes}}}

\item{node}{A character string. The quoted name of a node.}
}
\description{
Helper to turn query into a data expression
}
