% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_query_to_nodal_type.R
\name{query_to_expression}
\alias{query_to_expression}
\title{Helper to turn query into a data expression}
\usage{
query_to_expression(query, node)
}
\arguments{
\item{query}{A character string. An expression defining nodal types to
interrogate. An expression of the form
"Y[X=1]" asks for the value of Y when X is set to 1}

\item{node}{A character string. The quoted name of a node.}
}
\value{
A cleaned query expression
}
\description{
Helper to turn query into a data expression
}
\keyword{internal}
