% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_to_rows.R
\name{query_to_parameters}
\alias{query_to_parameters}
\title{Function to translate queries into parameter names}
\usage{
query_to_parameters(model, query)
}
\description{
Function to translate queries into parameter names
}
