% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_helpers.R
\name{query_to_parameters}
\alias{query_to_parameters}
\title{Function to translate queries into parameter names}
\usage{
query_to_parameters(model, query)
}
\arguments{
\item{model}{A model created by make_model()}

\item{query}{a query}
}
\description{
Function to translate queries into parameter names
}
\examples{

}
\keyword{internal}
