% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{reduce_nodal_types}
\alias{reduce_nodal_types}
\title{Restrict a model}
\usage{
reduce_nodal_types(dag, restrictions)
}
\arguments{
\item{dag}{a}

\item{restrictions}{list}
}
\value{
A DAG
}
\description{
Restrict a model
}
