% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{restrict_by_labels}
\alias{restrict_by_labels}
\title{Reduce nodal types using labels}
\usage{
restrict_by_labels(model, labels, keep = FALSE)
}
\arguments{
\item{model}{a model created by make_model()}

\item{labels}{A list of character vectors specifying nodal types to be kept or removed from the model.}

\item{keep}{Logical. If `FALSE`, removes and if `TRUE` keeps only causal types specified by \code{restriction}.}
}
\description{
Reduce nodal types using labels
}
