% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_restrictions.R
\name{restrict_by_labels}
\alias{restrict_by_labels}
\title{Reduce nodal types using labels}
\usage{
restrict_by_labels(model, labels, keep = FALSE)
}
\arguments{
\item{model}{a model created by make_model()}

\item{labels}{A list of character vectors specifying nodal types to be kept or removed from the model.}

\item{keep}{Logical. If `FALSE`, removes and if `TRUE` keeps only causal types specified by \code{restriction}.}
}
\description{
Reduce nodal types using labels
}
\examples{
model <- make_model('X->Y')
gbiqq:::restrict_by_labels(model, labels = list(X='0', Y = '00'))$parameters_df
gbiqq:::restrict_by_labels(model, labels = list(X='0', Y = '00'), keep = TRUE)$parameters_df
gbiqq:::restrict_by_labels(model, labels = list(X='0', Y = '?0'))$parameters_df
gbiqq:::restrict_by_labels(model, labels = list(X='0', Y = '?0'), keep = TRUE)$parameters_df

}
\seealso{
Other restrictions: 
\code{\link{restrict_by_query}()},
\code{\link{set_restrictions}()}
}
\concept{restrictions}
