% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_ambiguity_matrices.R
\name{reveal_data}
\alias{reveal_data}
\title{Reveal data}
\usage{
reveal_data(dag)
}
\arguments{
\item{dag}{A dag as created by \code{make_dag}}
}
\value{
revealed_data
}
\description{
Reveals data given types of exogenous variables
}
\examples{
# ADD_EXAMPLES_HERE
}
\keyword{internal}
