% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguities.R
\name{reveal_outcomes}
\alias{reveal_outcomes}
\title{Reveal outcomes}
\usage{
reveal_outcomes(model)
}
\arguments{
\item{model}{A model as created by \code{make_model}}
}
\value{
revealed_data
}
\description{
Reveal outcomes for all causal types. Calculated by sequentially calculating endogenous variables
}
\details{
\code{reveal_outcomes} starts off by creating types (via \code{\link{gbiq:::get_types}}). It then takes types of endogenous and reveals their outcome based on the value that their parents took. Exogenous variables outcomes correspond to their type.
}
\examples{
model <- make_model(add_edges(parent = "X", children = "Y"))
reveal_outcomes(model)

}
