% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguities.R
\name{reveal_outcomes}
\alias{reveal_outcomes}
\title{Reveal outcomes}
\usage{
reveal_outcomes(model, dos = NULL, node = NULL)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{dos}{A named list. Do actions defining variable values, e.g., \code{list(X = 0, M = 1)}.}
}
\value{
A data.frame object of revealed data.
}
\description{
Reveal outcomes for all causal types. Calculated by sequentially calculating endogenous variables.
If a do operator is applied to any variable then it takes the given value and all its descendants are generated accordingly.
}
\details{
\code{reveal_outcomes} starts off by creating types (via \code{\link{get_types}}). It then takes types of endogenous and reveals their outcome based on the value that their parents took. Exogenous variables outcomes correspond to their type.
}
\examples{
model <- make_model("X -> Y")
reveal_outcomes(model)

model <- make_model("X1->Y;X2->M;M->Y")
reveal_outcomes(model, dos = list(X1 = 1, M = 0))

}
