% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realise_outcomes.R
\name{reveal_outcomes}
\alias{reveal_outcomes}
\title{Reveal outcomes}
\usage{
reveal_outcomes(model, dos = NULL, node = NULL)
}
\description{
`r lifecycle::badge("deprecated")`

This function was deprecated because the name causes clashes with
DeclareDesign. Use realise_outcomes instead.
}
\keyword{internal}
