% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ambiguities_matrix.R
\name{set_ambiguities_matrix}
\alias{set_ambiguities_matrix}
\title{Set ambiguity matrix}
\usage{
set_ambiguities_matrix(model, A = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{A}{An ambiguities matrix}
}
\description{
Add an ambiguities matrix to a model
}
