% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_confounds.R
\name{set_confound}
\alias{set_confound}
\title{Set confound}
\usage{
set_confound(model, confound = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{confound}{A named list relating nodes to statements that identify causal types with which they are  confounded}
}
\description{
Adjust parameter matrix to allow confounding
}
\examples{
model <- make_model("A -> B -> C -> D; B -> D")
confound = list(X = "(D[A=., B=1, C=1]>D[A=., B=0, C=0])")
set_confound(model = model, confound = confound)
model <- make_model("X -> Y")
confound = list(X = "(Y[X=1]>Y[X=0])")
set_confound(model = model, confound = confound)
}
