% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_confounds_df.R
\name{set_confounds_df}
\alias{set_confounds_df}
\title{Set confounds dataframe}
\usage{
set_confounds_df(model, confounds_df = NULL, ...)
}
\arguments{
\item{model}{A model made by make_model}

\item{...}{arguments passed to make_confounds_df}
}
\description{
Adds matrix of confounded nodal types to model.
}
\examples{
model <- make_model("X -> M -> Y; X <-> M;  M <-> Y") \%>\%
   set_confounds_df()
get_confounds_df(model)


}
