% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_confounds_df.R
\name{set_confounds_df}
\alias{set_confounds_df}
\title{Set a confounds_df}
\usage{
set_confounds_df(model)
}
\arguments{
\item{model}{A model made by make_model}
}
\description{
Normally a confounds_df is added to a model whenever confounding is set.
The confounds_df can be manually provided however using set_sconfounds_df.
}
\examples{
model <- make_model('X -> Y') \%>\%
  set_confound(list('X <-> Y'), add_confounds_df = FALSE)
model$confounds_df
set_confounds_df(model)$confounds_df

# An example where a confounds dataframe needs to be updated manually
# Here a restriction applied after a confounding relation is set removes the confounding
model <- make_model('X -> Y') \%>\%
  set_confound(list(X = '(Y[X=1] > Y[X=0])')) \%>\%
  set_restrictions('(Y[X=1] > Y[X=0])')
model$confounds_df  # Incorrect
model <- set_confounds_df(model)
model$confounds_df  # Correct
# plot(model)
}
