% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_lambda}
\alias{set_lambda}
\title{Add a true prior distribution}
\usage{
set_lambda(model, lambda = NULL)
}
\arguments{
\item{model}{A model generated by \code{make_model}}

\item{n_draws}{Number of draws}
}
\description{
Add a vector of true parameters -- can be use to draw data, rather than true vector
}
