% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_parameter_matrix.R
\name{set_parameter_matrix}
\alias{set_parameter_matrix}
\title{Set parameter matrix}
\usage{
set_parameter_matrix(model, P = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{P}{A parameter matrix}
}
\description{
Add a parameter matrix to a model
}
