% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_parameters}
\alias{set_parameters}
\title{Add a true prior distribution}
\usage{
set_parameters(model, parameters = NULL, type = NULL)
}
\arguments{
\item{model}{A model generated by \code{make_model}}

\item{parameters}{parameter vector to add to model}

\item{type}{An element of ("flat", "prior_mean", "posterior_mean", "prior_draw", "posterior_draw"). Indicates type of parameters to set.}
}
\description{
Add a vector of true parameters -- can be use to draw data, rather than true vector
}
\examples{
model <- set_parameters(make_model("X -> Y"), type = "flat")

}
