% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_parameters}
\alias{set_parameters}
\title{Add a true parameter vector}
\usage{
set_parameters(model, parameters = NULL, type = "prior_mean", ...)
}
\arguments{
\item{model}{A model generated by \code{make_model}}

\item{parameters}{parameter vector to add to model}

\item{type}{An element of ("flat", "new", prior_mean", "posterior_mean", "prior_draw", "posterior_draw"). Indicates type of parameters to set.}
}
\description{
Add a vector of true parameters -- can be use to draw data, rather than true vector
}
\examples{
model <- set_parameters(make_model("X -> Y"),
          type = "flat")
# Use make_alpha to define specific parameters
model <- set_parameters(make_model("X -> Y"), type = "new",
                        alphas = list(Y = c(`(Y[X=1] > Y[X=0])` = 3)))
model$parameters
}
