% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{set_parameters}
\alias{set_parameters}
\title{Add a true parameter vector}
\usage{
set_parameters(model, parameters = NULL, type = "prior_mean", ...)
}
\arguments{
\item{model}{A model generated by \code{make_model}.}

\item{parameters}{A numeric vector. Parameters to add to model.}

\item{type}{A character string specifying type of parameters to set ("flat", "prior_mean", "posterior_mean", "prior_draw", "posterior_draw", "define). With type set to \code{define} use arguments to be passed to \code{make_priors}; otherwise \code{flat} sets equal probabilities on each nodal type in each parameter set; \code{prior_mean}, \code{prior_draw}, \code{posterior_mean}, \code{posterior_draw} take parameters as the means or as draws from the prior or posterior.}

\item{...}{Options passed onto \code{make_priors()}.}
}
\description{
Add a vector of "true" parameters to a model.
}
\examples{
model <- make_model("X -> Y")
get_parameters(model)
set_parameters(model, parameters = c(.25, .75, 1.25,.25, .25, .25)) \%>\% get_parameters
set_parameters(make_model("X -> Y"), type = "flat") \%>\% get_parameters

# Use make_alpha to define specific parameters using causal syntax
set_parameters(make_model("X -> Y"), type = 'define',
                        alphas = 1) \%>\% get_parameters
}
