% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_priors.R
\name{set_prior_distribution}
\alias{set_prior_distribution}
\title{Add prior distribution draws}
\usage{
set_prior_distribution(model, n_draws = 4000)
}
\arguments{
\item{model}{A model object generated by make_model().}

\item{n_draws}{A scalar. Number of draws.}
}
\description{
Add n_param x n_draws database of possible lambda draws to the model.
}
\examples{
model <- make_model("X -> Y") \%>\% set_prior_distribution(5)
model$prior_distribution

}
