% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_priors}
\alias{set_priors}
\title{Set prior distribution}
\usage{
set_priors(model, priors = NULL, prior_distribution = "uniform",
  alphas = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{priors}{Existing vector of Dirichlet hyperparameters}

\item{prior_distribution}{A character indicating the prior distribuiton. One of "jeffreys", "certainty", "uniform".}

\item{alphas}{Vector of hyperparameters of the dirichlet distribution to be passed to \code{make_priors}.}
}
\description{
Set a prior distribution. Use \code{make_alphas} to use causal types to indicate priors.
}
\examples{

model <- gbiqq::make_model("X -> Y")
set_priors(model = model,
           alphas = list(
             X = c(`X == 1` = 3),
             Y = c(`(Y[X=1] != Y[X=0])` = 3)))

}
