% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_priors}
\alias{set_priors}
\title{Set prior distribution}
\usage{
set_priors(model, lambda_priors, prior_distribution = NULL,
  alphas = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{lambda_priors}{Vector of Dirichlet hyperparameters}

\item{prior_distribution}{A character indicating the prior distribuiton}

\item{alphas}{the hyperparameters of the dirichlet distribution.}
}
\description{
Set prior distribution
}
