% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_priors}
\alias{set_priors}
\title{Set prior distribution}
\usage{
set_priors(model, lambda_priors = NULL, prior_distribution = "uniform",
  alphas = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{lambda_priors}{Existing vector of Dirichlet hyperparameters}

\item{prior_distribution}{A character indicating the prior distribuiton}

\item{alphas}{Vector of hyperparameters of the dirichlet distribution to be passed to \code{make_priors}.}
}
\description{
Set prior distribution
}
