% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_priors}
\alias{set_priors}
\title{Set prior distribution}
\usage{
set_priors(model, priors = NULL, prior_distribution = "uniform",
  alphas = NULL)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{priors}{A numeric vector. Dirichlet hyperparameters.}

\item{prior_distribution}{A character string. Indicates the prior distribution to be used. Can be 'uniform', 'jeffreys', or 'certainty'.}

\item{alphas}{A numeric vector. Hyperparameters of the Dirichlet distribution to be passed to \code{make_priors()}.}
}
\description{
Set prior distribution. Use \code{make_alphas} to use causal types to indicate priors.
}
\examples{

library(dplyr)
model <- make_model("X -> Y") \%>\%
  set_priors(
           alphas = list(
             X = c(`X == 1` = 3),
             Y = c(`(Y[X=1] != Y[X=0])` = 3)))
get_priors(model)

model <- make_model("X -> Y") \%>\%
set_priors(prior_distribution = "jeffreys")
get_priors(model)

model <- make_model("X -> Y") \%>\%
set_priors(1:6)
get_priors(model)
}
