% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_priors.R
\name{set_priors}
\alias{set_priors}
\title{Set prior distribution}
\usage{
set_priors(model, priors = NULL, distribution = NA, alphas = NA,
  node = NA, label = NA, statement = NA, confound = NA)
}
\arguments{
\item{model}{A model created with \code{make_model}}

\item{priors}{A optional vector of positive reals indicating priors over all parameters. These are interepreted as arguments for Dirichlet distributions---one for each parameter set. To see the structure of parameter sets examine model$parameters_df}

\item{distribution}{String (or list of strings) indicating a common prior distribution (uniform, jeffreys or certainty)}

\item{alphas}{Real positive numbers giving hyperparameters of the Dirichlet distribution}

\item{node}{A string (or list of strings) indicating nodes for which priors are to be altered}

\item{label}{String. Label for nodal type indicating nodal types for which priors are to be altered}

\item{statement}{A causal query (or list of queries) that determines nodal types for which priors are to be altered}

\item{confound}{A confound statement (or list of statements) that restricts nodal types for which priors are to be altered}
}
\description{
A flexible function to add priors to a model.
}
\details{
Four arguments govern *which* parameters should be altered. The default is "all" but this can be reduced by specifying

* \code{label} The label of a particular nodal type, written either in the form Y0000 or Y.Y0000

* \code{node}, which restricts for example to parameters associated with node "X"

* \code{statement}, which restricts for example to nodal types that satisfy the statment "Y[X=1] > Y[X=0]"

* \code{confound}, which restricts for example to nodal types that satisfy the statment "Y[X=1] > Y[X=0]"

Two arguments govern what values to apply:

* alphas is one or more non negative numbers and

* "distribution" indicates one of a common class: uniform, jeffreys, or "certain"

Any arguments entered as lists or vectors of size > 1 should be of the same length as each other.

For more examples and details see \code{make_priors}
}
\examples{

library(dplyr)
model <- make_model("X -> Y") \%>\%
  set_priors(alphas = 3)
get_priors(model)

model <- make_model("X -> Y") \%>\%
set_priors(distribution = "jeffreys")
get_priors(model)

model <- make_model("X -> Y") \%>\%
set_priors(1:6)
get_priors(model)
model <- make_model("X -> Y") \%>\%
set_priors(node = "Y", alphas = 2)
get_priors(model)
}
\seealso{
Other priors: \code{\link{get_priors}},
  \code{\link{make_priors_single}},
  \code{\link{make_priors}}
}
\concept{priors}
