% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{set_sampling_args}
\alias{set_sampling_args}
\title{set_sampling_args
From rstanarm (November 1st, 2019)}
\usage{
set_sampling_args(object, user_dots = list(), user_adapt_delta = NULL, ...)
}
\description{
Set the sampling arguments
}
