% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Generate full dataset, possibly from compact data}
\usage{
simulate_data(model, n = 1, data_events = NULL, parameters = NULL)
}
\arguments{
\item{model}{A model created by make_model()}

\item{n}{Number of observations}

\item{data_events}{A compact dataframe compatible with model}

\item{parameters}{A specific parameter vector, parameters, may be provided, otherwise parameters is drawn from priors}
}
\description{
Generate full dataset, possibly from compact data
}
\examples{
model <- make_model("X -> Y")
data_events <- draw_data_events(model = model, n = 4)
draw_data(model, data_events = data_events)
}
