% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Generate full dataset, possibly from compact data}
\usage{
simulate_data(model, n = 1, data_events = NULL, parameters = NULL,
  using = "parameters")
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{n}{An integer. Number of observations.}

\item{data_events}{A compact data frame compatible with \code{model}.}

\item{parameters}{A numeric vector. Values of parameters may be specified. By default, parameters is drawn from priors.}

\item{using}{String indicating whether to use "priors", "posteriors" or "parameters". Defaults to "parameters".}
}
\value{
A data frame of simulated data.
}
\description{
Generate full dataset, possibly from compact data
}
\examples{
model <- make_model("X -> Y")
data_events <- draw_data_events(model = model, n = 4)
simulate_data(model, data_events = data_events)
}
