% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Generate full dataset, possibly from compact data}
\usage{
simulate_data(model, n = 1, parameters = NULL, using = "parameters",
  w = NULL, P = NULL, A = NULL)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{n}{An integer. Number of observations.}

\item{parameters}{A numeric vector. Values of parameters may be specified. By default, parameters is drawn from priors.}

\item{using}{String indicating whether to use "priors", "posteriors" or "parameters". Defaults to "parameters".}

\item{w}{Vector of event probabilities can be provided directly. This is useful for speed for repeated data draws.}

\item{P}{A parameter matrix, can be used to generate w if w is not provided}

\item{A}{An ambiguity matrix, can be used to generate w if w is not provided}
}
\value{
A data frame of simulated data.
}
\description{
Generate full dataset, possibly from compact data
}
\examples{
model <- make_model("X -> Y")
 # Simplest behavior uses by default
 # the parameter vector contained in model,
 # which is flat by default:
simulate_data(model, n = 5)

# Simulate multiple datasets is fastest if w is provided
w <- draw_event_prob(model, using = "parameters")
replicate(5, simulate_data(model, n = 5, w = w))

}
