% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate_data is an alias for make_data}
\usage{
simulate_data(...)
}
\arguments{
\item{...}{arguments for make_model}
}
\description{
simulate_data is an alias for make_data
}
\examples{
simulate_data(make_model("X->Y"))
}
