% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_events.R
\name{simulate_events}
\alias{simulate_events}
\title{Draw compact data}
\usage{
simulate_events(
  model,
  n = 1,
  w = NULL,
  P = NULL,
  A = NULL,
  parameters = NULL,
  param_type = NULL,
  clean_params = TRUE,
  include_strategy = FALSE,
  ...
)
}
\arguments{
\item{model}{A model object generated by \code{make_model()}.}

\item{n}{An integer. Number of observations.}

\item{w}{A numeric  N(parameters) x 1 matrix of event probabilities with named rows.}

\item{P}{Parameter matrix. (Optional). May be provided to avoid repeated computation for simulations.}

\item{A}{Ambiguity matrix. (Optional). May be provided to avoid repeated computation for simulations.}

\item{parameters}{A numeric vector. Values of parameters may be specified. By default, it is drawn from priors.}

\item{param_type}{A character string specifying type of parameters to make ("flat", "prior_mean", "posterior_mean", "prior_draw", "posterior_draw", "define). With param_type set to \code{define} use arguments to be passed to \code{make_priors}; otherwise \code{flat} sets equal probabilities on each nodal type in each parameter set; \code{prior_mean}, \code{prior_draw}, \code{posterior_mean}, \code{posterior_draw} take parameters as the means or as draws from the prior or posterior.}

\item{clean_params}{Logical. Use to check parameters in case these are manually supplied.}

\item{include_strategy}{Logical. Whether to include a "strategy" vector. Defaults to FALSE. Strategy vector does not vary with full data but expected by some functions.}

\item{...}{Arguments to be passed to make_priors if param_type == \code{define}}
}
\value{
A data frame of events
}
\description{
Draw compact data
}
\examples{
model <- make_model("X -> Y")
simulate_events(model = model)
simulate_events(model = model, param_type = "prior_draw")
simulate_events(model = model, include_strategy = TRUE)

}
