% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summarise_distribution}
\alias{summarise_distribution}
\title{helper to compute mean and sd of a distribution data.frame}
\usage{
summarise_distribution(x)
}
\description{
helper to compute mean and sd of a distribution data.frame
}
