% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize Data}
\usage{
summarize_data(model, data)
}
\arguments{
\item{model}{A probabilistic causal model created by  \code{make_model()}}

\item{data}{A data.frame of variables that can take three values: 0, 1, and NA.}
}
\value{
Returns data events without strategies containing no observed data
}
\description{
Summarize Data
}
