% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_models.R
\name{summary.causal_model}
\alias{summary.causal_model}
\alias{print.summary.causal_model}
\title{Summarizing causal models}
\usage{
\method{summary}{causal_model}(object, ...)

\method{print}{summary.causal_model}(x, stanfit = FALSE, ...)
}
\arguments{
\item{object}{An object of \code{causal_model} class produced using
\code{make_model} or \code{update_model}.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of \code{summary.causal_model} class, usually a result of
a call to \code{summary.causal_model}.}

\item{stanfit}{Logical. Whether to include readable summary of
\code{stanfit} produced when updating a model via \code{update_model}.
Defaults to `FALSE`.}
}
\description{
summary method for class \code{causal_model}.
}
\details{
\code{print.summary.causal_model} reports DAG data frame, full specification of
nodal types and summary of model restrictions in addition to standard
\code{print.causal_model} output.
}
