% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.model_query}
\alias{summary.model_query}
\title{Summarizing model queries}
\usage{
\method{summary}{model_query}(object)
}
\arguments{
\item{object}{An object of \code{query_model} class produced using
\code{query_model}}
}
\value{
Returns the object of class \code{summary.model_query}
}
\description{
summary method for class "\code{model_query}".
}
