% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representing_dags.R
\name{translate_dagitty}
\alias{translate_dagitty}
\title{Puts your DAG into daggity syntax (useful for using their plotting functions)}
\usage{
translate_dagitty(dag)
}
\arguments{
\item{dag}{a dag from a model created by make_model()}
}
\value{
dagitty translation of DAG
}
\description{
Puts your DAG into daggity syntax (useful for using their plotting functions)
}
\examples{
\dontrun{
model <- make_model(
 add_edges(parent = "X",children = c("K","Y")),
 add_edges(parent = "K",children = "Y")
)

translate_dagitty(model$dag)
}

}
