% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dag.R
\name{translate_dagitty}
\alias{translate_dagitty}
\title{Puts your DAG into daggity syntax (useful for using their plotting functions)}
\usage{
translate_dagitty(model)
}
\arguments{
\item{model}{A model created by make_model()}
}
\value{
dagitty translation of DAG
}
\description{
If confounds are indicated (provided in \code{attr(model$P, "confounds")}), then these are represented as bidirectional arcs.
}
\examples{
\dontrun{
model <- make_model("X -> Y")
translate_dagitty(model)
}

}
