% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{trim_strategies}
\alias{trim_strategies}
\title{Trim strategies}
\usage{
trim_strategies(data_events)
}
\arguments{
\item{data_events}{A data.frame whose columns are \code{event}, \code{strategy} and \code{count}}
}
\value{
Returns data events with strategies (excluding  strategy families that contain no observed data)
}
\description{
Trim strategies
}
\examples{
model <- make_model("X -> Y")
data <- simulate_data(model, n = 10)
data[1,1] <- ""
summarize_data(model, data)
}
