% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{trim_strategies}
\alias{trim_strategies}
\title{trim_strategies}
\usage{
trim_strategies(model, data)
}
\arguments{
\item{model}{A probabilistic causal model created by make_model()}
}
\value{
Returns indices and ambiguity matrix
}
\description{
trim_strategies
}
