% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictions.R
\name{update_causal_types}
\alias{update_causal_types}
\title{Update causal types based on nodal types}
\usage{
update_causal_types(model)
}
\arguments{
\item{model}{A model object generated with \code{make_model}}
}
\description{
Update causal types based on nodal types
}
\examples{
update_causal_types(make_model("X->Y"))
}
