% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_restrictions.R
\name{update_causal_types}
\alias{update_causal_types}
\title{Update causal types based on nodal types}
\usage{
update_causal_types(model)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}
}
\value{
A \code{data.frame} containing updated causal types in a model
}
\description{
Update causal types based on nodal types
}
\examples{
CausalQueries:::update_causal_types(make_model('X->Y'))
}
\keyword{internal}
