% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{var_in_query}
\alias{var_in_query}
\title{List of nodes contained in query}
\usage{
var_in_query(model, query)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}

\item{query}{A character string. An expression defining nodal types to interrogate \code{\link{reveal_outcomes}}}
}
\description{
List of nodes contained in query
}
