% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{var_in_query}
\alias{var_in_query}
\title{List of nodes contained in query}
\usage{
var_in_query(model, query)
}
\arguments{
\item{model}{A \code{causal_model}. A model object generated by \code{\link{make_model}}.}

\item{query}{A character string. An expression defining nodal types to interrogate \code{\link{realise_outcomes}}. An expression of the form "Y[X=1]" asks for the value of Y when X is set to 1}
}
\value{
A vector indicating which variables are included in a query
}
\description{
List of nodes contained in query
}
\keyword{internal}
